% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ability.R
\name{ability}
\alias{ability}
\title{Ability estimation}
\usage{
ability(data, items, age = "age", key = NULL, metric = "dscore",
  full = FALSE, dec = 2, ...)
}
\arguments{
\item{data}{A data.frame containing columns names \code{items} with PASS/FAIL observations. 
Scores are coded numerically as \code{pass = 1} and \code{fail = 0}. 
Alternatively, \code{pass = TRUE} and \code{fail = FALSE} may be used. Additionally,
a column named \code{age} specifying decimal age in years.}

\item{items}{A character vector with item names in the chosen \code{lexicon}. 
The default is \code{names(scores)}.}

\item{age}{A character with the name of the age variable 
specifying decimal age in years. This information 
is used 1) to break up calculations into separate D-scores per age, 
and 2) to specify age-dependent priors.}

\item{key}{A data.frame of two columns containing the \code{items} names and a second column 
named \code{delta} containing the item difficulties estimated from the Rasch model in the 
prefferred metric/scale.}

\item{metric}{A string, either \code{"dscore"} or \code{"logit"}, 
signalling the metric in which the ability are estimated}

\item{full}{DOCUMENTATIONNEEDED}

\item{dec}{Number of decimals of the EAP estimates. Default is 2.}

\item{\dots}{Additional parameters passed down to \code{gettau()} (e.g., 
\code{lexicon} or \code{itembank}) and \code{adp()} (e.g., \code{mu} 
\code{sd} or \code{reference}).}
}
\value{
If \code{full} is \code{FALSE}: 
A vector of EAP estimates per age 
with \code{length(unique(age))} elements. If \code{full} is \code{TRUE}: 
A \code{list} of \code{length(unique(age))} elements with the
starting prior, posterior, quadrature grid and eap.
}
\description{
This function estimates the ability, a numerical score that measures
generic abilty, from PASS/FAIL observations (items).
}
\details{
By default, the algorithm produces a list containing the ful positior
and the expected a posteriori (EAP) estimate from \code{scores} using item 
difficulty as obtained from the item bank, and an age-dependent
prior of about twice the normal variation at the given age.
The method uses Bayes rule. See Bock and Mislevy (1982) for more 
details on the EAP and its variance. 

The mixing coefficients \code{mem.between} and \code{mem.within} are 
0-1 numbers (fractions) that regulate the influence 
of the previous posterior on the prior. 
There are three cases, depending on whether it's the 
first age and first item:
\describe{
\item{First age, first item:}{For the first item at the first age
\code{t}, the prior is set to normal with mean equal to the
median D-score at age \code{t} as calculated from \code{Dreference[,"mu"]} and 
where the standard deviation is equal to 5, about twice the normal variation in \code{Dreference}.}
\item{Later age, first item:}{For the first item at the second or later age \code{t}, the 
prior is set to \code{mem.between * post + (1 - mem.between) * adp()}, 
where \code{post} is the D-score posterior as calculated from the 
previous age, and where \code{adp()} is the age-dependent prior at
age \code{t}. The mixing coefficient \code{mem.between = 0} by default, which
implies that D-scores of the same child at different ages are calculated 
independently.}
\item{Any age, later item:}{For the second or further items at any 
given age \code{t}, the prior
is set to \code{mem.within * post + (1 - mem.within) * adp()}, where 
\code{post} is the D-score posterior as calculated from the 
previous item within age \code{t}, and 
where \code{adp()} is the age-dependent prior at age \code{t}. 
By default the mixing coefficient \code{mem.within = 1}, which
mimmicks repeated application of Bayesian rule to the current data 
using 'old' data as prior.}
}

param prior: The mean of the prior. If \code{mu = "gcdg"} (the default)
then \code{mu} is calculated from the Count model coded in 
\code{dscore:::count_mu_gcdg()}. Specify \code{mu = "reference"} in order
to take it from the age-dependent reference (default < 0.22).
}
\examples{
\dontrun{
data <- ddata::get_gcdg(study="Netherlands 1", adm=TRUE)  
data$age <- data$age/12    
items <- dmetric::prepare_items(study="Netherlands 1")$items
ability(data=data, items=items,lexicon="gcdg", itembank=gcdg_itembank)
}
}
\references{
Bock DD, Mislevy RJ (1982).  
Adaptive EAP Estimation of Ability in a Microcomputer Environment.
Applied Psychological Measurement, 6(4), 431-444.

Van Buuren S (2014). Growth charts of human development.
Stat Methods Med Res, 23(4), 346-368.
}
\seealso{
\code{\link{adp}}, \code{\link{gettau}}, 
\code{\link{itembank}}, \code{\link{posterior}},
\code{\link{Dreference}}
}
\author{
Stef van Buuren 2016
}
