% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adp.R
\name{adp}
\alias{adp}
\title{Age-dependent prior}
\usage{
adp(qp, mu, sd)
}
\arguments{
\item{qp}{Numeric vector of equally spaced quadrature points.
This vector should span the range of all D-score values. The default
(\code{qp = -10:100}) is suitable for age range 0-4 years.}

\item{mu}{Numeric scalar. The mean of the prior.}

\item{sd}{Numeric scalar. Standard deviation of the prior.}
}
\value{
A \code{vector} of \code{length(qp)} elements with 
the prior density estimate at each quadature point \code{qp}.
}
\description{
Returns the age-dependent prior N(mu, 5) at the 
specified quadrature points.
}
\note{
Use \code{qp = -10:80} to reproduce 0-2 year estimates in 
Van Buuren (2014).
}
\examples{
# define quadrature points for D-score, as used in Van Buuren 2014
qp <- -10:80

# calculate and plot three priors
plot(x = qp, y= adp(qp, 20, 5), type = "l", 
  main = "Priors at ages of 1, 12 and 24 months", 
  ylab = "Density", xlab = "D-score")
lines(x = qp, adp(qp, 40, 3), lty = 2)
lines(x = qp, adp(qp, 60, 6), lty = 3)
}
\references{
Van Buuren S (2014). Growth charts of human development.
Stat Methods Med Res, 23(4), 346-368.
}
\seealso{
\code{\link{dscore}}
}
