% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builtin_domaintable.R
\docType{data}
\name{builtin_domaintable}
\alias{builtin_domaintable}
\title{Developmental domain for items}
\format{
A \code{data.frame} with variables:\tabular{ll}{
   Name \tab Label \cr
   \code{set} \tab Item name, gsed lexicon \cr
   \code{domain} \tab name of the developmental domain \cr
   \code{item} \tab Item name, gsed lexicon \cr
   \code{weight} \tab Proportional weight in domain based on expert votes \cr
}
}
\usage{
builtin_domaintable
}
\description{
The built-in variable \code{builtin_domaintable} contains the domain and weight
of items for measuring early child development.
}
\details{
The \code{builtin_domaintable} is created by script
\code{data-raw/R/save_builtin_domaintable.R}.

Votings are collected by Gareth MacCray in surveys administered to several
subject matter experts. Each SME could attribute domains to an item, the
expert could distribute their vote to multiple domains if applicable. SME
votes were weighted to a proportional distribution of votes over the domains
for each item.
}
\author{
Compiled by Iris Eekhout using different sources
}
\keyword{datasets}
