% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builtin_translate.R
\docType{data}
\name{builtin_translate}
\alias{builtin_translate}
\title{A table to translate between different lexicons (naming schema)}
\format{
A \code{data.frame} with variables:\tabular{ll}{
   Name \tab Label \cr
   \code{phase1} \tab Item names, Phase 1 data \cr
   \code{phase2} \tab Item names, Phase 2 data \cr
   \code{gsed} \tab gsed lexion \cr
   \code{gsed2} \tab gto/gpa lexicon for LF/SF \cr
   \code{gsed3} \tab gl1/gs1 lexicon for LF/SF \cr
   \code{short1} \tab Short item name, phase 1 order \cr
   \code{short2} \tab Short item name, phase 2 order \cr
   \code{instrument} \tab Instrument code \cr
   \code{seq_phase1} \tab Phase 1 order \cr
   \code{seq_phase2} \tab Phase 2 order \cr
   \code{label} \tab Item label (English) \cr
}
}
\usage{
builtin_translate
}
\description{
The built-in variable \code{builtin_translate} contains a table for
translating among sets of item names into each other.
}
\details{
The \code{builtin_translate} is created by script
\code{data-raw/R/save_builtin_translate.R}.

Updates:
\itemize{
\item July 2025 - Tranferred from gsedread package
}
}
\author{
Compiled by Stef van Buuren
}
\keyword{datasets}
