% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_posterior.R
\name{calculate_posterior}
\alias{calculate_posterior}
\title{Calculate posterior of ability}
\usage{
calculate_posterior(scores, delta, age, qp = -10:100, mem.between = 0,
  mem.within = 1, ...)
}
\arguments{
\item{scores}{BLABLA}

\item{delta}{BLABLA}

\item{age}{BLABLA}

\item{qp}{A number vector of equally spaced quadrature points.
This vector should span the range of all D-score values, and 
have at least 80 elements. The default is 
\code{qp = -10:100}, which is suitable for age range 0-4 years.}

\item{mem.between}{Fraction of posterior from previous occasion 
relative to age-dependent prior. The 
value \code{mem.between = 0} (the default) means that 
no smoothing over age is performed, while a \code{mem.between = 1} 
corresponds to maximal smoothing over age. See details.}

\item{mem.within}{Fraction of posterior from previous item within
the same age  relative to age-dependent prior. The 
value \code{mem.within = 1} (the default) means that all items 
count equally in the posterior, while a \code{mem.within = 0} 
corresponds to counting only the last item. See details.}

\item{\dots}{Additional parameters passed down to BLABLA}
}
\description{
Calculate posterior of ability
}
