% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/category_size_exceeds.R
\name{category_size_exceeds}
\alias{category_size_exceeds}
\title{Select_if helpers}
\usage{
category_size_exceeds(x, k = 10)
}
\arguments{
\item{x}{Numeric vector. This argument can be left unspecified when the
function is called by \code{select_if()}.}

\item{k}{Minimum number per category needed.}
}
\value{
Logical
}
\description{
Helper functions for the \code{select_if()} function.
}
\examples{
library("ddata")
library("dplyr")
data <- ddata::gcdg \%>\%
 filter(study == "Netherlands 2") \%>\%
 select_(.dots = item_names("NL2"))
dim(data)

# take only columns with at least 100 observations per category
data <- data \%>\% select_if(category_size_exceeds, 100)
dim(data)
}

