% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_mu.R
\name{count_mu_gsed}
\alias{count_mu_gsed}
\title{Median of GSED references}
\usage{
count_mu_gsed(t)
}
\arguments{
\item{t}{Decimal age, numeric vector}
}
\value{
A vector of length \code{length(t)} with the median of the GCDG references.
}
\description{
Returns the age-interpolated median of the GSED2208 references
based on LGF & SF in GSED-BGD, GSED-PAK, GSED-TZA.
}
\details{
The interpolation consists of three age pieces:

Count model: <= 9MN: 21.3449 + 26.4916 t +  7.0251(t + 0.2)
Count model: > 9Mn & <= 3.5 YR: 14.69947 - 12.18636 t + 69.11675(t + 0.92)
Linear model: > 3.5 YRS: 61.40956 + 3.80904 t

The working range is 0-3.5 years. After the age of 3.5 years, the function
will increase at an arbitrary rate of 3.8 D-score points per year.
This function should be called for remodelled \code{key = "gsed"} created
in August 2022.
}
\note{
Internal function. Called by \code{dscore()}
}
\examples{
dscore:::count_mu_gsed(0:5)
}
\author{
Stef van Buuren, on behalf of GSED
}
