% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_mu.R
\name{count_mu_phase1_healthy}
\alias{count_mu_phase1_healthy}
\title{Median of phase1_healthy references}
\usage{
count_mu_phase1_healthy(t)
}
\arguments{
\item{t}{Decimal age, numeric vector}
}
\value{
A vector of length \code{length(t)} with the median of the GCDG references.
}
\description{
Returns the age-interpolated median of the phase1 references
based on LF & SF in GSED-BGD, GSED-PAK, GSED-TZA.
}
\details{
This function is intended to be called when \code{key = "gsed2212"},
\code{key = "gsed2208"} or \code{key = "293_0"}.
}
\note{
Internal function. Called by \code{dscore()}
}
\examples{
dscore:::count_mu_phase1_healthy(0:5)
}
\author{
Stef van Buuren, on behalf of GSED project
}
