% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daz.R
\name{daz}
\alias{daz}
\title{D-score standard deviation score: DAZ}
\usage{
daz(d, x, x.unit = "year", reference = dscore::Dreference, dec = 3)
}
\arguments{
\item{d}{Vector of D-scores, typically calculated by \code{dscore()}}

\item{x}{Vector of ages}

\item{x.unit}{Units given in \code{x} specified by 
\code{"year"}, \code{"month"} or \code{""day}. The}

\item{reference}{The LMS reference values. The default uses the 
built-in reference \code{dscore::Dreference} for Dutch children
published in Van Buuren (2014). The table should contain the columns
\code{nu}, \code{mu} and \code{sigma}, and at least one of the columns
\code{year}, \code{month} or \code{day}.}

\item{dec}{The number of decimals (default \code{dec = 3}).}
}
\value{
Vector with Z-scores with \code{length(d)} elements
}
\description{
DAZ stands for "Development - Age adjusted Z-score".
}
\references{
Cole TJ, Green PJ (1992). Smoothing reference centile curves: The LMS 
method and penalized likelihood. Statistics in Medicine, 11(10), 
1305-1319.
}
\seealso{
\code{\link{dscore}}
}

