% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddomain.R
\name{ddomain}
\alias{ddomain}
\title{Domain specific D-score}
\usage{
ddomain(
  data,
  domainset,
  domain = NULL,
  vote_weight = NULL,
  items = names(data),
  key = NULL,
  population = NULL,
  xname = "age",
  xunit = c("decimal", "days", "months"),
  itembank = NULL,
  ...
)
}
\arguments{
\item{data}{A \code{data.frame} or \code{matrix} with the data.
A row collects all observations made on a child on a set of
milestones administered at a given age. The function calculates
a D-score for each row. Different rows can correspond to different
children or ages.}

\item{domainset}{String. The name of the set of domains to use. See
\code{with(builtin_domaintable, table(set, domain))} for the domain names in each
set.}

\item{domain}{character vector of the name of the domain(s) for which to
compute the domain score. Per default all domains in the \code{domainset} are
used .}

\item{vote_weight}{minimum proportion of votes (weight) for a domain that an
item needs to have to count for that domain.}

\item{items}{A character vector containing names of items to be
included into the D-score calculation. Milestone scores are coded
numerically as \code{1} (pass) and \code{0} (fail). By default,
D-score calculation is done on all items found in the data
that have a difficulty parameter under the specified \code{key}.}

\item{key}{String. They key identifies 1) the difficulty estimates
pertaining to a particular Rasch model, and 2) the prior mean and standard
deviation of the prior distribution for calculating the D-score.
The default key \code{NULL} sets \code{key = "gsed2406"}.
View \code{builtin_keys} for an overview of the available keys.}

\item{population}{String. The name of the reference population to calculate
DAZ.
Use \code{with(builtin_references, table(key, population))} to see which
built-in references are available for \code{key - population} combinations.
If not specified, the function set the default population as
\code{builtin_keys$base_population[key == builtin_keys$key]}.}

\item{xname}{A string with the name of the age variable in
\code{data}. The default is \code{"age"}. Do not round age.}

\item{xunit}{A string specifying the unit in which age is measured
(either \code{"decimal"}, \code{"days"} or \code{"months"}).
The default \code{"decimal"} corresponds to decimal age in years.}

\item{itembank}{A \code{data.frame} with at least three columns named
\code{key}, \code{item} and \code{tau}. By default, the function uses
\code{dscore::builtin_itembank}. If you specify your own \code{itembank},
then you should also provide the relevant \code{transform} and \code{qp} arguments.}
}
\value{
The \code{ddomain()} function returns a list of \code{data.frame} objects with
each \code{nrow(data)} rows. The name of the list is the name of the domain.
The \code{data.frame} consists of the following columns:\tabular{ll}{
   Name \tab Label \cr
   \code{a} \tab Decimal age (years) \cr
   \code{n} \tab Number of items with valid (0/1) data \cr
   \code{p} \tab Percentage of passed milestones \cr
   \code{d} \tab D-score, mean of posterior distribution \cr
   \code{sem} \tab Standard error of measurement, standard deviation of the posterior \cr
   \code{daz} \tab D-score corrected for age, calculated in Z-scale (for metric \code{"dscore"}) \cr
}


The D-score in column \code{d} is a linear scale, with values usually ranging
from 0 to 100. The D-score is \code{NA} if age is missing or if age is lower
than -1/12. It is possible to calculate D-scores for cases with missing ages
by setting \code{prior_mean_NA} and \code{prior_sd_NA} to some reasonable value, e.g.,
\code{prior_mean_NA = 50} and \code{prior_sd_NA = 20}, for the sample at hand.

The SEM is a positive number that quantifies the uncertainty of the D-score.
It is \code{NA} if the D-score is \code{NA}.

The DAZ in column \code{daz} is a Z-score that corrects the D-score for age. It
is \code{NA} when there are no reference values for the given age, or when
the D-score is extremely unlikely to be valid at the given age.
}
\description{
Domain specific D-score
}
\examples{
sample <- dscore::gsample
colnames(sample) <- rename_vector(colnames(sample), lexin = "gsed2", lexout = "gsed3")
sample <- sample |> select(subjid, agedays, starts_with("gs1")) |>
 mutate(age = agedays / 365.25)
ddomain(sample, domainset = "partou")
ddomain(sample, domainset = "partou", domain = c("finemotor", "grossmotor"))
ddomain(sample, domainset = "partou", domain = c("language"))
}
\seealso{
\code{\link[=dscore]{dscore()}} \code{\link[=builtin_domaintable]{builtin_domaintable()}}
}
