% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{gcdg_itembank}
\alias{gcdg_itembank}
\title{Itembank from model 565_18}
\format{A \code{data.frame} with 565 rows and 6 variables:
\describe{
\item{lex_gcdg}{Item naming for GCDG project}
\item{instrument}{Instrument name}
\item{equate}{Equate group}
\item{domain}{Developmental domain}
\item{label}{Item label}
\item{tau}{Difficulty level, model 565_18}
}}
\usage{
gcdg_itembank
}
\description{
A data frame with administrative information per item: 
item name (in various systems), item labels and difficulty.
}
\details{
The itembank was calculated from 15 cohorts of the GCDG-study.
}
\examples{
head(gcdg_itembank)

}
\seealso{
\code{\link{dscore}}, \code{\link{gettau}}
}
\keyword{datasets}
