% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{gcdg_itembank_m}
\alias{gcdg_itembank_m}
\title{Expanded version of itembank including Mullen}
\format{An object of class \code{data.frame} with 728 rows and 6 columns.}
\usage{
gcdg_itembank_m
}
\description{
Expanded version of itembank including Mullen
}
\keyword{datasets}
