% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{gcdg_reference}
\alias{gcdg_reference}
\title{Age-conditional reference distribution of D-score}
\format{A \code{data.frame} with 121 rows and 19 variables:
\describe{
\item{day}{Age in days}
\item{week}{Age in weeks}
\item{year}{Decimal age in years}
\item{mu}{M-curve, median D-score, P50}
\item{sigma}{S-curve, spread expressed as coefficient of variation}
\item{nu}{L-curve, the lambda coefficient of the LMS model for skewness}
\item{P3}{Calculated third percentile}
\item{\dots}{And so on..}
}}
\usage{
gcdg_reference
}
\description{
A data frame containing the age-dependent distribution of the 
D-score for children aged 0-5 years. The distribution is modelled
by a normal distribution, and is equal for
both boys and girls. The LMS values can be used to graph 
reference charts and to calculate age-conditonal Z-scores, also 
known as \code{DAZ}.
}
\details{
The references were calculated from 15 cohorts of the GCDG-study.
}
\examples{
head(gcdg_reference)

}
\references{
Cole TJ, Green PJ (1992). Smoothing reference centile curves: The LMS 
method and penalized likelihood. Statistics in Medicine, 11(10), 
1305-1319.

Van Buuren S (2014). Growth charts of human development.
Stat Methods Med Res, 23(4), 346-368.
}
\seealso{
\code{\link{dscore}}, \code{\link{adp}}
}
\keyword{datasets}
