% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasch.R
\name{get_diff}
\alias{get_diff}
\title{Extract difficulty estimates}
\usage{
get_diff(object)
}
\arguments{
\item{object}{An object that inherits class \code{eRm}}
}
\value{
A named vector with the Rasch item difficulty estimates.
}
\description{
This function extracts the difficulty estimates from a fitted Rasch model.
}

