% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mu.R
\name{get_mu}
\alias{get_mu}
\title{Median D-score from the base population for a given key}
\usage{
get_mu(t, key, prior_mean_NA = NA_real_)
}
\arguments{
\item{t}{Decimal age, numeric vector}

\item{key}{Character, key of the reference population}

\item{prior_mean_NA}{Numeric, prior mean when age is missing}
}
\value{
A vector of length \code{length(t)} with the median of the default reference
population for the key.
}
\description{
Returns the age-interpolated median of the D-score of the default
reference for a given key.
}
\details{
Use \code{get_reference()} for more options.
}
