% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mu.R
\name{get_mu_gsed_cohorts}
\alias{get_mu_gsed_cohorts}
\title{Median D-score from GSED cohorts}
\usage{
get_mu_gsed_cohorts(
  t,
  key = "gsed2510",
  cohort = c("GSED-BGD", "GSED-BRA", "GSED-CHN", "GSED-CIV", "GSED-NLD", "GSED-PAK",
    "GSED-TZA")
)
}
\arguments{
\item{t}{Decimal age, numeric vector}

\item{key}{Character, key name}

\item{cohort}{Character, cohort name}
}
\value{
A vector of length \code{length(t)} with the median of the specified GSED cohort
references.
}
\description{
Returns the age-interpolated median of the GSED cohorts
based on LF & SF in seven GSED countries. This function is used
to set prior mean for poulations "GSED-BGD", "GSED-BRA", etc.
}
