% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_palette.R
\name{get_palette}
\alias{get_palette}
\title{Obtains color palettes for the Jamaica data}
\usage{
get_palette(palet = c("study", "country", "domain", "instrument", "wave",
  "item"), which_palettes = "jamaica_palettes", na_color = "grey")
}
\arguments{
\item{palet}{The name of the list component in the built-in 
color palette list \code{jamaica_palettes}. Currently implemented
are \code{"study"}, \code{"country"}, \code{"domain"}, 
\code{"instrument"}, \code{"wave"} and \code{"item"}.}

\item{which_palettes}{A named list containing palettes. The default is the
built-in object \code{jamaica_palettes}.}

\item{na_color}{The color that is return if not found. The default is 
\code{"grey"}.}
}
\value{
The selected color palette. If not found, it returns the single 
 color specified by \code{na_color}.
}
\description{
This function returns one of built-in color palettes from a 
list of palettes.
}
\examples{
 get_palette("domain")
}

