% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gettau.R
\name{gettau}
\alias{gettau}
\title{Obtain difficulty parameters from item bank}
\usage{
gettau(items, itembank = dscore::itembank, lexicon = "ghap", check = TRUE,
  ...)
}
\arguments{
\item{items}{A character vector with item names in the 
chosen lexicon.}

\item{itembank}{A \code{data.frame} that contains the item names 
(in various lexicons), the item label, the item difficulty parameter
\code{tau}. Lexicon column names start with \code{"lex."}. 
The default \code{dscore::itembank} contains the 
difficulty levels as published in Van Buuren (2014).}

\item{lexicon}{A character string indicating the column in 
\code{itembank} used to match item names. It must be one of 
the lexicon columns, e.g., \code{dutch1983}, 
\code{dutch1996}, \code{dutch2005}, \code{smocc} 
\code{ghap} or \code{gcdg}. The default is \code{lexicon = "ghap"}.}

\item{check}{Logical that indicates whether the lexicon name
should be checked. The default is \code{TRUE}.}

\item{\dots}{Additional arguments (ignored).}
}
\value{
A named vector with the difficulty estimate per item with
\code{length(items)} elements, or \code{NULL} if items are 
not found.
}
\description{
Searches the item bank for matching items, and returns the 
difficulty estimates. Matching is done by item name. Comparisons
are done in lower case.
}
\examples{
# difficulty levels in default GHAP lexicon
gettau(items = c("GSFIXEYE", "GSMARMR"))

# difficulty levels of same items in the SMOCC lexicon
gettau(items = c("v1430", "v1432"), lexicon = "smocc")

}
\seealso{
\code{\link{itembank}}, \code{\link{dscore}}
}
\author{
Stef van Buuren 2016
}
