% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsample.R
\docType{data}
\name{gsample}
\alias{gsample}
\title{Sample of 10 children from the GSED Phase 1 study}
\format{
A \code{data.frame} with 10 rows and 295 variables:\tabular{ll}{
   Name \tab Label \cr
   \code{id} \tab Integer, child ID \cr
   \code{agedays} \tab Integer, age in days \cr
   \code{gpalac001} \tab Integer, Cry when hungry...: 1 = yes, 0 = no, NA = not administered \cr
   \code{gpalac002} \tab Integer, Look at/focus...: 1 = yes, 0 = no, NA = not administered \cr
   \code{...} \tab and so on.. \cr
}


There are 138 \code{gpa} items (item \code{gpamoc008} (clench fists) removed) from GSED SF and
and 155 \code{gto} items from GSED LF.
}
\usage{
gsample
}
\description{
A demo dataset with developmental scores at the item level for
10 random children from the GSED Phase 1 data.
}
\examples{
head(gsample)
}
\seealso{
\code{\link[=dscore]{dscore()}}
}
\keyword{datasets}
