% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{jamaica_palettes}
\alias{jamaica_palettes}
\title{Color palettes for the Jamaica data}
\format{A named \code{list} with 5 elements.}
\usage{
jamaica_palettes
}
\description{
A list containing five color palettes for the Jamaica data.
}
\keyword{datasets}

