% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexicon.R
\name{lexicon}
\alias{lexicon}
\title{Find lexicon item names}
\usage{
lexicon(lexicon = "gsed", itembank = dscore::builtin_itembank,
  na.omit = TRUE)
}
\arguments{
\item{lexicon}{Item naming scheme. The built-in itembank supports
lexicons \code{"gsed"} (default), \code{"gcdg"}, \code{"ghap"}, 
\code{"dutch1983"}, \code{"dutch1996"}, \code{"dutch2005"} 
and \code{"smocc"}. See details.}

\item{itembank}{A \code{data.frame} that contains the item names 
(in various lexicons), the item label, and item difficulty 
parameters \code{tau} under various models. 
Lexicon column names start with \code{"lex_"}. 
The function uses the \code{dscore::builtin_itembank} object by 
default.}

\item{na.omit}{Logical signalling where \code{NA} should be omitted.
The default is \code{na.omit = TRUE}.}
}
\value{
A character vector with all items recognized under the 
lexicon-itembank combination. May be of length zero if the 
lexicon or itembank is not found.
}
\description{
The function returns the item names recognized under a specified
lexicon in a given itembank.
}
\examples{
lexicon("smocc")
}
\seealso{
\code{\link{dscore}}
}
