% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subfuncDscorev7.R
\name{normalize}
\alias{normalize}
\title{Normalize distribution}
\usage{
normalize(d, theta)
}
\arguments{
\item{d}{A vector with \code{length(theta)} elements representing
the unscaled density at each quadrature point.}

\item{theta}{Vector of equally spaced quadrature points.}
}
\value{
A \code{vector} of \code{length(d)} elements with 
the prior density estimate at each quadature point.
}
\description{
Normalizes the distribution so that the total mass equals 1.
}

