% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{popsdemo}
\alias{popsdemo}
\title{Project On Preterm and Small for gestational age infants (POPS)}
\format{A \code{tbl_df} with 100 rows (25 children times 4 time points) and 67 variables:
\describe{
\item{patid}{Patient ID}
\item{gender}{1=male, 1 = female}
\item{gestionalage}{Gestational age in fractional weeks}
\item{moment}{Sequence number}
\item{age}{Age in days}
\item{occ}{Measurement occasion (1-4)}
\item{daycor}{Corrected age in days}
\item{Dscore}{Calculated D-score}
\item{popsdscoresds}{Corrected age-conditional standard deviation score of D-score}
\item{dead}{0 = alive, 1 = dead}
\item{Fixateseyes}{Fixates eyes: 0 = yes, 1 = no}
\item{\dots}{and so on..}
}}
\source{
\url{https://www.tno.nl/nl/aandachtsgebieden/gezond-leven/prevention-work-health/voor-tijdens-en-na-de-zwangerschap/pops-langlopend-onderzoek-onder-prematuren-en-baby-s-met-laag-geboortegewicht/}
\code{\link{dscore}}
}
\usage{
popsdemo
}
\description{
A dataset with developmental scores at the item level from a subset
of 25 children from the POPS cohort.
}
\note{
Note that the PASS/FAIL items in the \code{pops} data are coded
in the `wrong` way: 0 = PASS, 1 = FAIL.
}
\examples{
head(popsdemo)

}
\keyword{datasets}
