% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dscore.R
\name{posterior}
\alias{posterior}
\title{Calculate posterior for one item given score, difficulty and prior}
\usage{
posterior(score, tau, prior, theta)
}
\arguments{
\item{score}{A scalar value between 0 and m}

\item{tau}{A vector of length \code{m} with uncentralized thresholds from the Rasch model}

\item{prior}{vector of prior values on quadrature points \code{theta}}

\item{theta}{vector of equally spaced quadrature points}
}
\value{
A vector of length \code{length(prior)}
}
\description{
Calculate posterior for one item given score, difficulty and prior
}
\details{
This function assumes that the difficulties have been estimated by 
a polytomous Rasch model, e.g. by \code{RUMM2030} or 
\code{sirt::rasch.pairwise()}, and transformed onto the 
appropriate scale. The response vector takes on values
\code{0:m}. The binary Rasch model assumes \code{m = 1}. The D-score
calculation assumes binary scores. 

Note: This function is intended to be used internally and 
does not check input arguments. Use \code{dscore()} to estimate
D-score from data.
}
\author{
Stef van Buuren 2016
}
\references{
Andrich D. Rasch Models for Measurement. Newbury Park: 
Sage Publications; 1988.

RUMM Laboratories. RUMM 2030. 
Rasch Unidimensional Measurement Models. Perth: 2015.
}

