% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsample.R
\docType{data}
\name{sample_lf}
\alias{sample_lf}
\title{Sample of 10 children from gto (LF)}
\format{
A \code{data.frame} with 10 rows and 157 variables:\tabular{ll}{
   Name \tab Label \cr
   \code{subjid} \tab Integer, child ID \cr
   \code{agedays} \tab Integer, age in days \cr
   \code{lf001} \tab Integer, ...: 1 = yes, 0 = no, NA = not administered \cr
   \code{lf002} \tab Integer, ...: 1 = yes, 0 = no, NA = not administered \cr
   \code{...} \tab and so on.. \cr
}


Sample data for 155 \code{gto} items from GSED SF
}
\usage{
sample_lf
}
\description{
A demo dataset with developmental scores at the item level for
10 random children from the GSED Phase 1 data.
}
\examples{
head(sample_lf)
}
\seealso{
\code{\link[=dscore]{dscore()}}
}
\keyword{datasets}
