% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsample.R
\docType{data}
\name{sample_sf}
\alias{sample_sf}
\title{Sample of 10 children from gpa (SF)}
\format{
A \code{data.frame} with 10 rows and 141 variables:\tabular{ll}{
   Name \tab Label \cr
   \code{subjid} \tab Integer, child ID \cr
   \code{agedays} \tab Integer, age in days \cr
   \code{sf001} \tab Integer, Cry when hungry...: 1 = yes, 0 = no, NA = not administered \cr
   \code{sf002} \tab Integer, Look at/focus...: 1 = yes, 0 = no, NA = not administered \cr
   \code{...} \tab and so on.. \cr
}


Sample data for 139 \code{gpa} items from GSED SF
}
\usage{
sample_sf
}
\description{
A demo dataset with developmental scores at the item level for
10 random children from the GSED Phase 1 data.
}
\examples{
head(sample_sf)
}
\seealso{
\code{\link[=dscore]{dscore()}}
}
\keyword{datasets}
