% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triple.R
\docType{data}
\name{triple}
\alias{triple}
\title{Sample of 50 children measured with three instruments}
\format{
A \code{data.frame} with 50 rows and 559 variables:\tabular{ll}{
   Name \tab Label \cr
   \code{id} \tab Integer, child ID \cr
   \code{age} \tab Numeric, age in decimal years \cr
   \code{agedays} \tab Integer, age in days \cr
   \code{gs1sec001} \tab Integer, SF001 Does your child smile? \cr
   \code{gs1moc002} \tab Integer, SF002 When lying on his/her back, ... \cr
   \code{...} \tab and so on.. \cr
}


The dataset contains 138 items from GSED SF (\code{gs1}),
(item \code{gs1moc028} was skipped), 155 items from GSED LF (\code{gl1}),
and 263 (out of 326) items from BSID-III (\code{by3}).
}
\usage{
triple
}
\description{
An example dataset with developmental scores at the item level for
50 random children from the GSED Validation Study (Cavellera et al, 2023).
Each child has measurements from GSED SF (\code{gs1}), GSED LF (\code{gl1}) and
BSID-III (\code{by3}).
}
\examples{
# calculate D-score from all instruments
ds_all <- dscore(triple)
head(ds_all)
# calculate D-score from only GSED SF items
ds_sf <- dscore(triple, items = get_itemnames(instrument = "gs1"))
head(ds_sf)
}
\references{
Cavallera et al. (2023). Protocol for validation of the Global
Scales for Early Development (GSED) for children under 3 years of
age in seven countries. BMJ Open, 13(1), e062562.
DOI: 10.1136/bmjopen-2022-062562.
\url{https://bmjopen.bmj.com/content/13/1/e062562}

World Health Organization (WHO) (2023). Global Scales for Early
Development (GSED) V1.0: Technical Report. Geneva: World Health
Organization.
\url{https://www.who.int/publications/i/item/WHO-MSD-GSED-package-v1.0-2023.1}
}
\seealso{
\code{\link[=dscore]{dscore()}}
}
\keyword{datasets}
