% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daz.R
\name{zad}
\alias{zad}
\title{Inverse D-score standard deviation score}
\usage{
zad(z, x = as.numeric(names(z)), x.unit = "year",
  ref = dscore::Dreference, dec = 3)
}
\arguments{
\item{z}{Vector of standard deviation scores (DAZ)}

\item{x}{Vector of ages. The default is to take age from 
\code{names(z)}.}

\item{x.unit}{Units given in \code{x} specified by 
\code{"year"}, \code{"month"} or \code{"day"}. The default is 
\code{"year"}.}

\item{ref}{The LMS reference values. The default uses the 
built-in reference \code{dscore::Dreference} for Dutch children
published in Van Buuren (2014). The table should contain the columns
\code{nu}, \code{mu} and \code{sigma}, and at least one of the columns
\code{year}, \code{month} or \code{day}.}

\item{dec}{The number of decimals (default \code{dec = 3}).}
}
\value{
Names vector with D-scores with \code{length(z)} elements
}
\description{
This function is the inverse of \code{daz()}: Given age and 
the Z-score, it finds the raw D-score.
}
\examples{
# population median at ages 0.5, 1 and 2 years
zad(z = rep(0, 3), x = c(0.5, 1, 2))

# percentiles of D-score reference
g <- expand.grid(age = seq(0.1, 2, 0.1), p = c(0.1,0.5,0.9))
d <- zad(z = qnorm(g$p), x = g$age)
matplot(x = matrix(g$age, ncol = 3), y = matrix(d, ncol = 3), type = "l",
lty = 1, col = "blue", xlab = "Age (years)", ylab = "D-score")
}
\references{
Cole TJ, Green PJ (1992). Smoothing reference centile curves: The LMS 
method and penalized likelihood. Statistics in Medicine, 11(10), 
1305-1319.
}
\seealso{
\code{\link{dscore}}, \code{\link{daz}}
}

