% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dscore.R
\name{adp}
\alias{adp}
\title{Age-dependent prior}
\usage{
adp(age, theta, mu = NULL, sd = 5, reference = dscore::Dreference, ...)
}
\arguments{
\item{age}{Numeric, single value}

\item{theta}{A number vector of equally spaced quadrature points}

\item{mu}{The mean of the prior. If \code{is.null(mu)} (the default) the prior is taken from the age-dependent reference.}

\item{sd}{Standard deviation of the prior. The default is 5.}

\item{reference}{the LMS reference values. The default uses the 
built-in reference \code{dscore::Dreference} for Dutch children
published in Van Buuren (2014).}

\item{\dots}{Additional parameters (ignored)}
}
\value{
A \code{vector} of \code{length(theta)} elements with 
the prior density estimate at each quadature point \code{theta}.
}
\description{
Returns the age-dependent prior N(mu, 5) at the 
specified quadrature points.
}
\examples{
# define quadrature points for D-score
theta <- -10:80

# calculate and plot three priors
plot(x = theta, y= adp(1/12, theta), type = "l", 
  main = "Priors at ages of 1, 12 and 24 months", 
  ylab = "Density", xlab = "D-score")
lines(x = theta, adp(1, theta), lty = 2)
lines(x = theta, adp(2, theta), lty = 3)
}
\references{
Van Buuren S (2014). Growth charts of human development.
Stat Methods Med Res, 23(4), 346-368.
}

