% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adp.R, R/dscore.R
\name{adp}
\alias{adp}
\alias{adp}
\title{Age-dependent prior}
\usage{
adp(age, qp = -10:100, mu = "dutch", sd = 5,
  reference = dscore::Dreference, ...)

adp(age, qp = -10:100, mu = "dutch", sd = 5,
  reference = dscore::Dreference, ...)
}
\arguments{
\item{age}{Age in years. Numeric, single value. If a vector, only the 
first value will be used.}

\item{qp}{A number vector of equally spaced quadrature points.
This vector should span the range of all D-score values, and 
have at least 80 elements. The default is 
\code{qp = -10:100}, which is suitable for age range 0-4 years.}

\item{mu}{The mean of the prior. If \code{mu = "dutch"} (the default)
then \code{mu} is calculated from the Count model coded in 
\code{dscore:::count_mu_dutch()}. Specify \code{mu = "reference"} in order
to take it from the age-dependent reference (default < 0.22).}

\item{sd}{Standard deviation of the prior. The default is 5.}

\item{reference}{the LMS reference values. The default uses the 
built-in reference \code{dscore::Dreference} for Dutch children
published in Van Buuren (2014).}

\item{\dots}{Additional parameters (ignored)}

\item{age}{Age in years. Numeric, single value. If a vector, only the 
first value will be used.}

\item{qp}{A number vector of equally spaced quadrature points.
This vector should span the range of all D-score values, and 
have at least 80 elements. The default is 
\code{qp = -10:100}, which is suitable for age range 0-4 years.}

\item{mu}{The mean of the prior. If \code{mu = "dutch"} (the default)
then \code{mu} is calculated from the Count model coded in 
\code{dscore:::count_mu_dutch()}. Specify \code{mu = "reference"} in order
to take it from the age-dependent reference (default < 0.22).}

\item{sd}{Standard deviation of the prior. The default is 5.}

\item{reference}{the LMS reference values. The default uses the 
built-in reference \code{dscore::Dreference} for Dutch children
published in Van Buuren (2014).}

\item{\dots}{Additional parameters (ignored)}
}
\value{
A \code{vector} of \code{length(qp)} elements with 
the prior density estimate at each quadature point \code{qp}.

A \code{vector} of \code{length(qp)} elements with 
the prior density estimate at each quadature point \code{qp}.
}
\description{
Returns the age-dependent prior N(mu, 5) at the 
specified quadrature points.

Returns the age-dependent prior N(mu, 5) at the 
specified quadrature points.
}
\note{
Use \code{qp = -10:80} to reproduce 0-2 year estimates in 
Van Buuren (2014).

Use \code{qp = -10:80} to reproduce 0-2 year estimates in 
Van Buuren (2014).
}
\examples{
# define quadrature points for D-score, as used in Van Buuren 2014
qp <- -10:80

# calculate and plot three priors
plot(x = qp, y= adp(1/12, qp), type = "l", 
  main = "Priors at ages of 1, 12 and 24 months", 
  ylab = "Density", xlab = "D-score")
lines(x = qp, adp(1, qp), lty = 2)
lines(x = qp, adp(2, qp), lty = 3)
# define quadrature points for D-score, as used in Van Buuren 2014
qp <- -10:80

# calculate and plot three priors
plot(x = qp, y= adp(1/12, qp), type = "l", 
  main = "Priors at ages of 1, 12 and 24 months", 
  ylab = "Density", xlab = "D-score")
lines(x = qp, adp(1, qp), lty = 2)
lines(x = qp, adp(2, qp), lty = 3)
}
\references{
Van Buuren S (2014). Growth charts of human development.
Stat Methods Med Res, 23(4), 346-368.

Van Buuren S (2014). Growth charts of human development.
Stat Methods Med Res, 23(4), 346-368.
}
\seealso{
\code{\link{dscore}}

\code{\link{dscore}}
}
