% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builtin_itembank.R
\docType{data}
\name{builtin_itembank}
\alias{builtin_itembank}
\title{Built-in itembank}
\format{A \code{data.frame} with 112 rows and 13 variables:
\describe{
\item{lex_dutch1996}{Item naming, VWO1996 version}
\item{lex_dutch2005}{Item naming, VWO2005 version}
\item{lex_dutch1983}{Item naming, VWO1983 version}
\item{lex_smocc}{Item naming, original SMOCC variable names}
\item{lex_ghap}{Item naming used by GHAP platform of BMGF}
\item{lex_gcdg}{Item naming for GCDG project}
\item{lex_gsed}{Item naming for GSED project}
\item{occ}{Occasion number}
\item{labelNL}{Label (Dutch)}
\item{labelEN}{Label (English)}
\item{tau}{Difficulty level}
\item{m}{ }
\item{LR}{Left-right item scores}
\item{hot}{Item used to estimate D-score? (1=y, 0=no)}
}}
\usage{
builtin_itembank
}
\description{
A data frame with administrative information per item: 
item name (in various systems), item labels and difficulty.
}
\examples{
head(builtin_itembank)

}
\references{
Van Buuren S (2014). Growth charts of human development.
Stat Methods Med Res, 23(4), 346-368.
}
\seealso{
\code{\link{dscore}}, \code{\link{gettau}}
}
\keyword{datasets}
