% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builtin_itemtable.R
\docType{data}
\name{builtin_itemtable}
\alias{builtin_itemtable}
\title{Global Scale for Early Development - itemtable}
\format{A \code{data.frame} with five columns:
\describe{
\item{\code{item}}{Item name, in GSED lexicon}
\item{\code{instrument}}{Instrument name (3-letter code)}
\item{\code{domain}}{Developmental domain (2-letter code)}
\item{\code{mode}}{Mode of administration (1-letter code)}
\item{\code{number}}{Item number (3-number code)}
\item{\code{equate}}{Equate group}
\item{\code{label}}{Item label}
}}
\usage{
builtin_itemtable
}
\description{
The built-in variable named \code{itemtable}
contains descriptions of all items found in the \code{gsed}
data.
}
\details{
Data are collected by the members of the Global Scale for Early
Development (GSED) group.
The \code{itemtable} is created by \code{\\data-raw\\R\\import_itemtable.R}.
}
\author{
Compiled by Stef van Buuren
}
\keyword{datasets}
