% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builtin_itemtable.R
\docType{data}
\name{builtin_itemtable}
\alias{builtin_itemtable}
\title{Global Scale for Early Development - itemtable}
\format{
A \code{data.frame} with variables:\tabular{ll}{
   Name \tab Label \cr
   \code{item} \tab Item name, gsed lexicon \cr
   \code{equate} \tab Equate group \cr
   \code{label} \tab Label (English) \cr
}
}
\usage{
builtin_itemtable
}
\description{
The built-in variable named \code{builtin_itemtable}
contains descriptions of all items found in the \code{gsed}
data.
}
\details{
Data are collected by the members of the Global Scales for Early
Development (GSED) group.
The \code{itemtable} is created by \verb{\\\\data-raw\\\\R\\\\save_builtin_itemtable.R}.

Last update:
\itemize{
\item May 30, 2022 - added gto (LF) and gpa (SF) items
\item June 1, 2022 - added seven gsd items
\item Nov 24, 2022 - Added instruments gs1, gs2
\item Dec 01, 2022 - Labels of gto replaced by correct order. This change invalidates
any analyses done on LF done after May 30, 2022 !!!
\item Dec 05, 2022 - Redefines gs1 and instrument for Phase 2, removes gs2 (139)
Adds gl1 (Long Form Phase 2 items 155)
\item Jan 05, 2023 - Adds 55 items from GSED-HH
}
}
\author{
Compiled by Stef van Buuren
}
\keyword{datasets}
