% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builtin_itemtable.R
\docType{data}
\name{builtin_itemtable}
\alias{builtin_itemtable}
\title{Global Scale for Early Development - itemtable}
\format{
A \code{data.frame} with variables:\tabular{ll}{
   Name \tab Label \cr
   \code{item} \tab Item name, gsed lexicon \cr
   \code{equate} \tab Equate group \cr
   \code{label} \tab Label (English) \cr
}
}
\usage{
builtin_itemtable
}
\description{
The built-in variable named \code{builtin_itemtable}
contains descriptions of all items found in the \code{gsed}
data.
}
\details{
Data are collected by the members of the Global Scale for Early
Development (GSED) group.
The \code{itemtable} is created by \verb{\\\\data-raw\\\\R\\\\save_builtin_itemtable.R}.

Last update: April 24, 2020
}
\author{
Compiled by Stef van Buuren
}
\keyword{datasets}
