% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builtin_keys.R
\docType{data}
\name{builtin_keys}
\alias{builtin_keys}
\title{D-score builtin_keys table}
\format{
A \code{data.frame} with variables:\tabular{ll}{
   Name \tab Label \cr
   \code{key} \tab String indicating a specific Rasch model (the key) \cr
   \code{n_items} \tab Number of items in the key \cr
   \code{n_instruments} \tab Number of instruments in the key \cr
   \code{intercept} \tab Intercept to convert logit into D-score \cr
   \code{slope} \tab Slope to convert logit into D-score \cr
   \code{deprecated} \tab Logical. Is key deprecated? \cr
}
}
\usage{
builtin_keys
}
\description{
A key represents a set of Rasch models that have the same item difficulty
estimates (tau). The key is used to calculate D-scores. D-scores can
only be compared when calculated with the same key. The current
recommendation for new projects is to choose key \code{gsed2212}.
}
\details{
Some case are depecreated and should not be used, other than for
replication. The \code{deprecated} column indicates if a key is deprecated.
}
\note{
20240609 SvB: Added \code{builtin_keys} table by \verb{data-raw\\data\\R\\save_builtin_keys.R}
}
\keyword{datasets}
