% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builtin_references.R
\docType{data}
\name{builtin_references}
\alias{builtin_references}
\title{Age-conditional reference distribution of D-score}
\format{
A \code{data.frame} with 18 variables:\tabular{ll}{
   Name \tab Label \cr
   \code{pop} \tab Population: \code{"dutch"}, \code{"gcdg"}, \code{"phase1"}, \code{"phase1_healthy"}, \cr
}


\if{html}{\out{<div class="sourceCode">}}\preformatted{     `"dutch_gsed2212"` |
}\if{html}{\out{</div>}}

| \code{age} | Decimal age in years |
| \code{mu}  | M-curve, median D-score, P50 |
| \code{sigma} | S-curve, spread expressed as coefficient of variation |
| \code{nu}    | L-curve, the lambda coefficient of the LMS/BCT model for skewness |
| \code{tau}   | Kurtosis parameter in the BCT model |
| \code{P3}    | P3 percentile |
| \code{P10}   | P10 percentile |
| \code{P25}   | P25 percentile |
| \code{P50}   | P50 percentile |
| \code{P75}   | P75 percentile |
| \code{P90}   | P90 percentile |
| \code{P97}   | P97 percentile |
| \code{SDM2}  | -2SD centile |
| \code{SDM1}  | -1SD centile |
| \code{SD0}   | 0SD centile, median |
| \code{SDP1}  | +1SD centile |
| \code{SDP2}  | +2SD centile |
}
\usage{
builtin_references
}
\description{
A data frame containing the age-dependent distribution of the
D-score for children aged 0-5 years. The distribution is modelled
after the LMS distribution (Cole & Green, 1992) or BCT model
(Stasinopoulos & Rigby, 2022) and is equal for
both boys and girls. The LMS/BCT values can be used to graph
reference charts and to calculate age-conditional Z-scores, also
known as DAZ.
}
\details{
The \code{"dutch"} references were calculated from the SMOCC data, and cover
age range 0-2.5 years (van Buuren, 2014).
The \code{"gcdg"} references were calculated from the 15 cohorts of the
GCDG-study, and cover age range 0-5 years (Weber, 2019).
The \code{"phase1"} references were calculated from the GSED Phase 1 validation
data (GSED-BGD, GSED-PAK, GSED-TZA) cover age range 2w-3.5 years. The
age range 3.5-5 yrs is linearly extrapolated and are only indicative.
The \code{"phase1_healthy"} references were calculated from the GSED Phase 1 validation
data (GSED-BGD, GSED-PAK, GSED-TZA) using a subset of children with
healthy development.
The \code{"dutch_gsed2212"} references were calculated from Dutch data using
the \code{gsed2212} key. This is a temporary name, and will be deprecated.
}
\examples{
head(builtin_references)
}
\references{
Cole TJ, Green PJ (1992). Smoothing reference centile curves: The LMS
method and penalized likelihood. Statistics in Medicine, 11(10),
1305-1319.

Van Buuren S (2014). Growth charts of human development.
Stat Methods Med Res, 23(4), 346-368.
\url{https://stefvanbuuren.name/publication/van-buuren-2014-gc/}

Weber AM, Rubio-Codina M, Walker SP, van Buuren S, Eekhout I,
Grantham-McGregor S, Caridad Araujo M, Chang SM, Fernald LCH,
Hamadani JD, Hanlon A, Karam SM, Lozoff B, Ratsifandrihamanana L,
Richter L, Black MM (2019). The D-score: a metric for interpreting
the early development of infants and toddlers across global settings.
BMJ Global Health, BMJ Global Health 4: e001724.
\url{https://gh.bmj.com/content/bmjgh/4/6/e001724.full.pdf}

Stasinopoulos M, Rigby R (2022). gamlss.dist: Distributions for
Generalized Additive Models for Location Scale and Shape,
R package version 6.0-3,
\url{https://CRAN.R-project.org/package=gamlss.dist}
}
\seealso{
\code{\link[=dscore]{dscore()}}
}
\keyword{datasets}
