% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_mu.R
\name{count_mu}
\alias{count_mu}
\title{Median D-score from the default references for the given key}
\usage{
count_mu(t, key)
}
\arguments{
\item{t}{Decimal age, numeric vector}

\item{key}{Character, key of the reference population}
}
\value{
A vector of length \code{length(t)} with the median of the default reference
population for the key.
}
\description{
Returns the age-interpolated median of the D-score of the default
reference for a given key.
}
\details{
Do not use this function if you want the median D-score for a specific
reference.
}
