% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_mu.R
\name{count_mu_dutch}
\alias{count_mu_dutch}
\title{Median of Dutch references}
\usage{
count_mu_dutch(t)
}
\arguments{
\item{t}{Decimal age, numeric vector}
}
\value{
A vector of length \code{length(t)} with the median of the Dutch references.
}
\description{
Returns the age-interpolated median of the Dutch references (van Buuren 2014).
The working range is 0-3 years. This function should
be called when the \code{key = "dutch"}.
}
\note{
Internal function. Called by \code{dscore()}
}
\examples{
dscore:::count_mu_dutch(0:2)
}
