% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mu.R
\name{count_mu_preliminary_standards}
\alias{count_mu_preliminary_standards}
\title{Median of preliminary_standards}
\usage{
count_mu_preliminary_standards(t)
}
\arguments{
\item{t}{Decimal age, numeric vector}
}
\value{
A vector of length \code{length(t)} with the median of the GCDG references.
}
\description{
Returns the age-interpolated median of the preliminary_standards
based on LF & SF in GSED-BGD, GSED-PAK, GSED-TZA. This function is used
to set prior mean under key \code{"gsed2406"}.
}
\note{
Internal function. Called by \code{dscore()}
}
\examples{
dscore:::count_mu_preliminary_standards(0:5)
}
\author{
Stef van Buuren, on behalf of GSED project
}
