% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daz.R
\name{daz}
\alias{daz}
\title{D-score standard deviation score: DAZ}
\usage{
daz(d, x = as.numeric(names(d)), x.unit = "year",
  ref = dscore::Dreference, dec = 3)
}
\arguments{
\item{d}{Vector of D-scores, typically calculated by \code{dscore()}}

\item{x}{Vector of ages. The default is to take age from 
\code{names(d)}.}

\item{x.unit}{Units given in \code{x} specified by 
\code{"year"}, \code{"month"} or \code{"day"}. The default is 
\code{"year"}.}

\item{ref}{The LMS reference values. The default uses the 
built-in reference \code{dscore::Dreference} for Dutch children
published in Van Buuren (2014). The table should contain the columns
\code{nu}, \code{mu} and \code{sigma}, and at least one of the columns
\code{year}, \code{month} or \code{day}.}

\item{dec}{The number of decimals (default \code{dec = 3}).}
}
\value{
Named vector with Z-scores with \code{length(d)} elements
}
\description{
DAZ stands for "Development - Age adjusted Z-score". 
The DAZ represents a child's D-score after adjusting 
for age by an external age-conditional reference.
}
\examples{
# preterms, uncorrected age, all Z-scores are fairly low
daz1 <- daz(d = pops$Dscore, x = pops$age/365.25)
daz2 <- daz(d = pops$Dscore, x = pops$age, x.unit = "day")
plot(daz1, daz2)
abline(0, 1, lty = 2)

# preterms, corrected age, most Z-scores are between -2 and +2 SD
daz3 <- daz(d = pops$Dscore, x = pops$daycor, x.unit = "day")
plot(daz2, daz3, xlab = "DAZ uncorrected age", ylab = "DAZ corrected age")
abline(0, 1, lty = 2, h = c(2, 0, -2), v = c(2, 0, -2))
}
\references{
Cole TJ, Green PJ (1992). Smoothing reference centile curves: The LMS 
method and penalized likelihood. Statistics in Medicine, 11(10), 
1305-1319.
}
\seealso{
\code{\link{dscore}}, \code{\link{daz}}
}

