% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddomain.R
\name{ddomain}
\alias{ddomain}
\title{Domain specific D-score}
\usage{
ddomain(
  data,
  set,
  domain = NULL,
  vote_weight = NULL,
  items = names(data),
  key = NULL,
  population = NULL,
  xname = "age",
  xunit = c("decimal", "days", "months"),
  prepend = NULL,
  itembank = NULL,
  metric = c("dscore", "logit"),
  prior_mean = NULL,
  prior_mean_NA = NULL,
  prior_sd = NULL,
  prior_sd_NA = NULL,
  transform = NULL,
  qp = NULL,
  dec = c(2L, 3L),
  relevance = c(-Inf, Inf),
  algorithm = c("current", "1.8.7"),
  verbose = FALSE
)
}
\arguments{
\item{data}{A \code{data.frame} or \code{matrix} with the data.
A row collects all observations made on a child on a set of
milestones administered at a given age. The function calculates
a D-score for each row. Different rows can correspond to different
children or ages.}

\item{set}{String. The name of the set of domains to use. See
\code{with(builtin_domaintable, table(set, domain))} for the domain names in each
set.}

\item{domain}{character vector of the name of the domain(s) for which to
compute the domain score. Per default all domains in the \code{set} are
used .}

\item{vote_weight}{minimum proportion of votes (weight) for a domain that an
item needs to have to count for that domain.}

\item{items}{A character vector containing names of items to be
included into the D-score calculation. Milestone scores are coded
numerically as \code{1} (pass) and \code{0} (fail). By default,
D-score calculation is done on all items found in the data
that have a difficulty parameter under the specified \code{key}.}

\item{key}{String. They key identifies 1) the difficulty estimates
pertaining to a particular Rasch model, and 2) the prior mean and standard
deviation of the prior distribution for calculating the D-score.
The default key \code{NULL} sets \code{key = "gsed2510"}.
View \code{builtin_keys} for an overview of the available keys.}

\item{population}{String. The name of the reference population to calculate
DAZ.
Use \code{with(builtin_references, table(key, population))} to see which
built-in references are available for \code{key - population} combinations.
If not specified, the function set the default population as
\code{builtin_keys$base_population[key == builtin_keys$key]}.}

\item{xname}{A string with the name of the age variable in
\code{data}. The default is \code{"age"}. Do not round age.}

\item{xunit}{A string specifying the unit in which age is measured
(either \code{"decimal"}, \code{"days"} or \code{"months"}).
The default \code{"decimal"} corresponds to decimal age in years.}

\item{prepend}{Character vector with column names in \code{data} that will
be prepended to the returned data frame. This is useful for copying
columns from data into the result, e.g., for matching.}

\item{itembank}{A \code{data.frame} with at least three columns named
\code{key}, \code{item} and \code{tau}. By default, the function uses
\code{dscore::builtin_itembank}. If you specify your own \code{itembank},
then you should also provide the relevant \code{transform} and \code{qp} arguments.}

\item{metric}{A string, either \code{"dscore"} (default) or
\code{"logit"}, signalling the metric in which ability is estimated.
\code{daz} is not calculated for the logit scale.}

\item{prior_mean}{\code{NULL} (default), a string, a numeric scalar, or
a numeric vector with  \code{nrow(data)} elements. The default value
\code{NULL} will consult the \code{base_population} field in \code{builtin_keys},
and use the corresponding median of that reference as prior mean for
the D-score. The string should refer to a column name in \code{data}
that contains user-supplied values of the prior mean for each observation.
A numeric scalar will be expanded to all observations. A numeric vector
will be used as is.}

\item{prior_mean_NA}{\code{NULL} (default) or a scalar numeric, representing
the prior mean for observations with missing ages. By default, D-scores
with missing ages will we \code{NA}. We suggest setting
\code{prior_mean_NA = 50} as a reasonable choice for samples between 0-3
years. The argument is ignored if \code{prior_mean} is specified per
observation, which gives you full control of priors for observations
with missing ages.}

\item{prior_sd}{\code{NULL} (default), a string, a numeric scalar, or
a numeric vector with \code{nrow(data)} elements. The default (\code{NULL})
uses a value of 5 for all ages. The string should refer to a column
name in \code{data} that contains user-supplied values of the prior sd
for each observation. A numeric scalar will be expanded to all
observations. A numeric vector will be used as is.}

\item{prior_sd_NA}{\code{NULL} (default) or a scalar numeric, representing
the prior sd for observations with missing ages. By default, D-scores
with missing ages will we \code{NA}. We suggest setting
\code{prior_sd_NA = 20} as a reasonable choice for samples between 0-3
years. The argument is ignored if \code{prior_sd} is specified per
observation, which gives you full control of priors for observations
with missing ages.}

\item{transform}{Numeric vector, length 2, containing the intercept
and slope of the linear transform from the logit scale into the
the D-score scale. The default (\code{NULL}) searches \code{builtin_keys}
for intercept and slope values.}

\item{qp}{Numeric vector of equally spaced quadrature points.
This vector should span the range of all D-score or logit values.
The default (\code{NULL}) creates \code{seq(from, to, by)} searching the
arguments from \code{builtin_keys}.}

\item{dec}{A vector of two integers specifying the number of
decimals for rounding the D-score and DAZ, respectively.
The default is \code{dec = c(2L, 3L)}.}

\item{relevance}{A numeric vector of length with the lower and
upper bounds of the relevance interval. The procedure calculates
a dynamic EAP for each item. If the difficulty level (tau) of the
next item is outside the relevance interval around EAP, the procedure
ignore the score on the item. The default is \code{c(-Inf, +Inf)} does not
ignore scores.}

\item{algorithm}{Computational method, for backward compatibility.
Either \code{"current"} (default) or \code{"1.8.7"} (deprecated).}

\item{verbose}{Logical. Print settings.}
}
\value{
The \code{ddomain()} function returns a list of \code{data.frame} objects with
each \code{nrow(data)} rows. The name of the list is the name of the domain.
The \code{data.frame} consists of the following columns:\tabular{ll}{
   Name \tab Label \cr
   \code{a} \tab Decimal age (years) \cr
   \code{n} \tab Number of items with valid (0/1) data \cr
   \code{p} \tab Percentage of passed milestones \cr
   \code{d} \tab D-score, mean of posterior distribution \cr
   \code{sem} \tab Standard error of measurement, standard deviation of the posterior \cr
   \code{daz} \tab D-score corrected for age, calculated in Z-scale (for metric \code{"dscore"}) \cr
}


The D-score in column \code{d} is a linear scale, with values usually ranging
from 0 to 100. The D-score is \code{NA} if age is missing or if age is lower
than -1/12. It is possible to calculate D-scores for cases with missing ages
by setting \code{prior_mean_NA} and \code{prior_sd_NA} to some reasonable value, e.g.,
\code{prior_mean_NA = 50} and \code{prior_sd_NA = 20}, for the sample at hand.

The SEM is a positive number that quantifies the uncertainty of the D-score.
It is \code{NA} if the D-score is \code{NA}.

The DAZ in column \code{daz} is a Z-score that corrects the D-score for age. It
is \code{NA} when there are no reference values for the given age, or when
the D-score is extremely unlikely to be valid at the given age.
}
\description{
Domain specific D-score
}
\examples{
sample <- dscore::gsample
colnames(sample) <- dscore::rename_vector(colnames(sample), lexin = "gsed2", lexout = "gsed3")
sample <- sample |> dplyr::select(subjid, agedays, starts_with("gs1")) |>
 dplyr::mutate(age = agedays / 365.25)
ddomain(sample, set = "GFCLS")
ddomain(sample, set = "GFCLS", domain = c("finemotor", "grossmotor"))
ddomain(sample, set = "GFCLS", domain = c("language"))
}
\seealso{
\code{\link[=dscore]{dscore()}} \code{\link[=builtin_domaintable]{builtin_domaintable()}}
}
