% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dscore-package.R
\docType{package}
\name{dscore-package}
\alias{dscore-package}
\title{dscore: D-score for Child Development}
\description{
The \code{dscore} package implements several tools needed to 
calculate the D-score, a numerical score that measures
generic development in children.
}
\note{
Development of this package was kindly supported under the Healthy
Birth, Growth and Development knowledge integration (HBGDki)
program of the Bill & Melinda Gates Foundation.
}
\section{User functions}{

The available functions are:
\tabular{ll}{
  \code{\link{dscore}}  \tab Estimate D-score and DAZ\cr
  \code{\link{dscore_posterior}}  \tab Calculate full posterior\cr
  \code{\link{daz}}     \tab Transform to age-adjusted standardized D-score\cr
  \code{\link{zad}}     \tab Inverse of \code{\link{daz}}\cr
  \code{\link{get_tau}}         \tab Get difficulty parameters from item bank\cr
  \code{\link{get_reference}}   \tab Get D-score reference\cr
  \code{\link{get_itemnames}}   \tab Extract item names \cr
  \code{\link{get_itemtable}}   \tab Get a subset from the itemtable\cr
  \code{\link{get_labels}}      \tab Get labels for items \cr
  \code{\link{order_itemnames}} \tab Sort item names  \cr
  \code{\link{sort_itemnames}}  \tab Sort item names  \cr
  \code{\link{rename_gcdg_gsed}}\tab Rename gcdg lexicon to gsed lexicon\cr
  }
}

\section{Built-in data}{

The package contains the following built-in data:
\tabular{ll}{
\code{\link{builtin_itembank}} \tab A \code{data.frame} containing 
the difficulty estimates of items according to final Rasch models.\cr
\code{\link{builtin_itemtable}} \tab A \code{data.frame} containing 
names and descriptions of items from 22 instruments.\cr
\code{\link{builtin_references}} \tab A \code{data.frame} with LMS 
reference values used to transform from D-score to DAZ, 
DAZ to D-score, and to calculate reference charts of normal development.\cr
\code{\link{milestones}} \tab A small demo dataset with PASS/FAIL 
responses from 27 preterms, measured at various ages between birth 
and 2.5 years.\cr
}
}

\references{
Van Buuren S (2014). Growth charts of human development.
Stat Methods Med Res, 23(4), 346-368.
}
