% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dscore-package.R
\docType{package}
\name{dscore-package}
\alias{dscore-package}
\title{dscore: A package for calculating D-scores}
\description{
The \code{dscore} package provides two important functions:
\code{dscore} and \code{daz}.
}
\note{
Development of this package was kindly suported by the HBGDki
program of the Bill & Melinda Gates Foundation
}
\section{\code{dscore} functions}{

The \code{dscore()} function estimates the D-score, 
a numerical score that measures
generic development in children, from PASS/FAIL observations 
on developental milestones (items). 
Item names are matched by name against the built-in item 
bank in \code{itembank}.
}

\section{\code{daz} functions}{

The \code{daz()} function calculated the age-adjusted standard
deviation score (Z-score, or DAZ) from the estimated D-score.
Analyzing DAZ instead of the D-score simplified modelling 
D-score trajectories over time. The \code{zad()} performs
the inverse operation, from DAZ to D-score.
}

\section{External data}{

The calculations requires three sets of external data:
\describe{
\item{data}{The PASS/FAIL responses on a appropriate set
of developmental items. These data are typically supplied by 
the user. For illustration, the package contain a built-in 
small demo dataset from the POPS study in \code{dscore::pops}.}
\item{itembank}{A table that contains the difficulty of each 
item, as estimated under the unidimensional Rasch 1-PL model. By 
default, calculations use the built-in \code{dscore::itembank}.}
\item{reference}{A table with LMS reference values used to transform to D-score to DAZ, or DAZ to D-score. By default, calculations
use the built-in references \code{dscore::Dreference}}
}
}
\references{
Van Buuren S (2014). Growth charts of human development.
Stat Methods Med Res, 23(4), 346-368.
}
\seealso{
\code{\link{dscore}}, \code{\link{itembank}}, 
\code{\link{Dreference}}, \code{\link{daz}}
}

