% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subfuncDscorev7.R
\name{dscore}
\alias{dscore}
\title{D-score calculation}
\usage{
dscore(scores, names = names(scores), lexicon = "ID.smock", age = NULL,
  EAP = TRUE, ...)
}
\arguments{
\item{scores}{A vector containing PASS/FAIL observations 
of a child on one or more developmental milestones. Scores are coded 
as numerically as \code{pass = 1} and \code{fail = 0}. 
Alternatively, \code{pass = TRUE} and \code{fail = FALSE} may be used.}

\item{names}{A character vector with item names in the chosen \code{lexicon}. 
The default is \code{names(responses)}.}

\item{lexicon}{A character string indicating the set of variable names against 
which \code{items} are matched. It may be either \code{ID.VWO1996}, 
\code{ID.VWO2005}, \code{ID.VWO1983} or \code{ID.smock} (the default).}

\item{age}{Numeric, decimal age of the child in years. This information 
is used to specify the age-dependent prior of about twice the normal 
variation.}

\item{EAP}{A logical indicating to return the mean of the posterior (\code{EAP = TRUE}, the default) or the full posterior (\code{EAP = FALSE}).}

\item{\dots}{Parameters passed down to \code{posterior()}, for example, 
to set quadrature points, the prior, or the calculation method.}
}
\value{
Either a single value (if \code{EAP == TRUE}), or a vector 
with the density estimate at each quadature point (if \code{EAP == FALSE}).
}
\description{
This function estimates the D-score, a numerical score that measures
generic development in children, from PASS/FAIL observations 
on developental
milestones. Variable names are matched by name against the built-in item 
bank in \code{itembank}. 
Variables that do not match, or for which the score is 
\code{NA}, are ignored.
}
\details{
By default, the algorithm produces the expected a 
posteriori (EAP) estimate from \code{scores} using item 
difficulty as obtained from the item bank, and - 
if age is given - an age-dependent
prior of about twice the normal variation at the given age.
The method uses Bayes rule. The function can also return the 
full posterior instead of the EAP.
}

