% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dscore.R
\name{dscore}
\alias{dscore}
\title{D-score estimation}
\usage{
dscore(scores, items = names(scores), ages, qp = -10:80, mem.between = 0,
  mem.within = 1, full = FALSE, dec = 2, ...)
}
\arguments{
\item{scores}{A vector containing PASS/FAIL observations 
of a child on one or more developmental milestones. Scores are coded 
numerically as \code{pass = 1} and \code{fail = 0}. 
Alternatively, \code{pass = TRUE} and \code{fail = FALSE} may be used. Alternatively, \code{scores} can be a \code{data.frame} 
containing columns named \code{items}, \code{scores} and \code{ages}.}

\item{items}{A character vector with item names in the chosen \code{lexicon}. 
The default is \code{names(scores)}.}

\item{ages}{Numeric vector with \code{length(scores)} elements, 
specifying decimal age in years. This information 
is used 1) to break up calculations into separate D-scores per age, 
and 2) to specify age-dependent priors.}

\item{qp}{A number vector of equally spaced quadrature points.
This vector should span the range of all D-score values, and 
have at least 80 elements.
The default \code{qp = -10:80} is suitable for children 
aged 0-2 years.}

\item{mem.between}{Fraction of posterior from previous occasion 
relative to age-dependent prior. The 
value \code{mem.between = 0} (the default) means that 
no smoothing over age is performed, while a \code{mem.between = 1} 
corresponds to maximal smoothing over age. See details.}

\item{mem.within}{Fraction of posterior from previous item within
the same age  relative to age-dependent prior. The 
value \code{mem.within = 1} (the default) means that all items 
count equally in the posterior, while a \code{mem.within = 0} 
corresponds to counting only the last item. See details.}

\item{full}{A logical indicating to return the mean of the posterior (\code{full = FALSE}, the default) or the full posterior 
(\code{full = TRUE}).}

\item{dec}{Number of decimals of the EAP estimates. Default is 2.}

\item{\dots}{Additional parameters passed down to \code{gettau()} (e.g., 
\code{lexicon} or \code{itembank}) and \code{adp()} (e.g., \code{mu} 
\code{sd} or \code{reference}).}
}
\value{
If \code{full} is \code{FALSE}: 
A vector of EAP estimates per age 
with \code{length(unique(age))} elements. If \code{full} is \code{TRUE}: 
A \code{list} of \code{length(unique(age))} elements with the
starting prior, posterior, quadrature grid and eap.
}
\description{
This function estimates the D-score, a numerical score that measures
generic development in children, from PASS/FAIL observations 
on developental milestones (items). 
Item names are matched by name against the built-in item 
bank in \code{itembank}. 
Items that do not match, or for which the score is 
\code{NA}, are ignored.
}
\details{
By default, the algorithm produces the expected a 
posteriori (EAP) estimate from \code{scores} using item 
difficulty as obtained from the item bank, and an age-dependent
prior of about twice the normal variation at the given age.
The method uses Bayes rule. See Bock and Mislevy (1982) for more 
details on the EAP and its variance. 
Optionally, the \code{dscore()} function returns the 
full posterior (instead of the EAP) if the argument 
\code{full = TRUE} is specified.

The mixing coefficients \code{mem.between} and \code{mem.within} are 
0-1 numbers (fractions) that regulate the influence 
of the previous posterior on the prior. 
There are three cases, depending on whether it's the 
first age and first item:
\describe{
\item{First age, first item:}{For the first item at the first age
\code{t}, the prior is set to normal with mean equal to the
median D-score at age \code{t} as calculated from \code{Dreference[,"mu"]} and 
where the standard deviation is equal to 5, about twice the normal variation in \code{Dreference}.}
\item{Later age, first item:}{For the first item at the second or later age \code{t}, the 
prior is set to \code{mem.between * post + (1 - mem.between) * adp()}, 
where \code{post} is the D-score posterior as calculated from the 
previous age, and where \code{adp()} is the age-dependent prior at
age \code{t}. The mixing coefficient \code{mem.between = 0} by default, which
implies that D-scores of the same child at different ages are calculated 
independently.}
\item{Any age, later item:}{For the second or further items at any 
given age \code{t}, the prior
is set to \code{mem.within * post + (1 - mem.within) * adp()}, where 
\code{post} is the D-score posterior as calculated from the 
previous item within age \code{t}, and 
where \code{adp()} is the age-dependent prior at age \code{t}. 
By default the mixing coefficient \code{mem.within = 1}, which
mimmicks repeated application of Bayesian rule to the current data 
using 'old' data as prior.}
}
}
\examples{
# GSFIXEYE: Fixate eyes
# GSRSPCH:  Reacts to speech (M; can ask parents)
# GSMLEG :  Same amount of movement in both legs
items <- c("GSFIXEYE", "GSRSPCH", "GSMLEG")
gettau(items)
age <- round(rep(21/365.25, 3), 4)  # age 21 days
dscore(c(1, 0, 0), items, age)

# two time points, one additional (overlapping) item
items <- c(items, items[3])
age <- round(c(age, 42/365.25), 4)  # add age 42 days
dscore(c(1, 0, 0, 1), items, age)

# save full posterior
fp <- dscore(c(1, 1, 0, 1), items, age, full = TRUE)
plot(fp[[1]]$qp, fp[[1]]$posterior, type = "l",
xlab = "D-score", ylab = "Density", 
main = "Age 21 days: 2 PASS, FAIL at GSMLEG")
lines(fp[[1]]$qp, fp[[1]]$start, lty = 2)

# hardly any difference between prior and posterior 
# because PASS score is uninformative at age 42 days
plot(fp[[2]]$qp, fp[[2]]$posterior, type = "l",
xlab = "D-score", ylab = "Density", main = "Age 42 days: PASS at GSMLEG")
lines(fp[[2]]$qp, fp[[2]]$start, lty = 2)

# However a FAIL score signals substantial delay at age 42 days
fp <- dscore(c(1, 1, 0, 0), items, age, full = TRUE)
plot(fp[[2]]$qp, fp[[2]]$posterior, type = "l",
xlab = "D-score", ylab = "Density", main = "Age 42 days: FAIL at GSMLEG")
lines(fp[[2]]$qp, fp[[2]]$start, lty = 2)

}
\author{
Stef van Buuren 2016
}
\references{
Bock DD, Mislevy RJ (1982).  
Adaptive EAP Estimation of Ability in a Microcomputer Environment.
Applied Psychological Measurement, 6(4), 431-444.

Van Buuren S (2014). Growth charts of human development.
Stat Methods Med Res, 23(4), 346-368.
}
\seealso{
\code{\link{adp}}, \code{\link{gettau}}, 
\code{\link{itembank}}, \code{\link{posterior}},
\code{\link{Dreference}}
}

