% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dscore-package.R, R/dscore.R
\docType{package}
\name{dscore}
\alias{dscore}
\alias{dscore-package}
\alias{dscore_posterior}
\title{dscore: D-score for Child Development}
\usage{
dscore(
  data,
  items = names(data),
  xname = "age",
  xunit = c("decimal", "days", "months"),
  prepend = NULL,
  key = NULL,
  itembank = dscore::builtin_itembank,
  metric = c("dscore", "logit"),
  prior_mean = NULL,
  prior_sd = NULL,
  transform = NULL,
  qp = NULL,
  population = NULL,
  dec = c(2L, 3L),
  relevance = c(-Inf, Inf),
  algorithm = c("current", "1.8.7")
)

dscore_posterior(
  data,
  items = names(data),
  xname = "age",
  xunit = c("decimal", "days", "months"),
  prepend = NULL,
  key = NULL,
  itembank = dscore::builtin_itembank,
  metric = c("dscore", "logit"),
  prior_mean = NULL,
  prior_sd = NULL,
  transform = NULL,
  qp = NULL,
  population = NULL,
  dec = c(2L, 3L),
  relevance = c(-Inf, Inf),
  algorithm = c("current", "1.8.7")
)
}
\arguments{
\item{data}{A \code{data.frame} with the data.
A row collects all observations made on a child on a set of
milestones administered at a given age. The function calculates
a D-score for each row. Different rows correspond to different
children or different ages.}

\item{items}{A character vector containing names of items to be
included into the D-score calculation. Milestone scores are coded
numerically as \code{1} (pass) and \code{0} (fail). By default,
D-score calculation is done on all items found in the data
that have a difficulty parameter under the specified \code{key}.}

\item{xname}{A string with the name of the age variable in
\code{data}. The default is \code{"age"}. Do not round age.}

\item{xunit}{A string specifying the unit in which age is measured
(either \code{"decimal"}, \code{"days"} or \code{"months"}).
The default \code{"decimal"} corresponds to decimal age in years.}

\item{prepend}{Character vector with column names in \code{data} that will
be prepended to the returned data frame. This is useful for copying
columns from data into the result, e.g., for matching.}

\item{key}{A string that selects a subset in the itembank that
makes up the key, the set of difficulty estimates from a fitted
Rasch model.
The built-in keys are:  \code{"gsed2212"} (default), \code{"gsed2208"} (deprecated),
\code{"gsed2206"} (deprecated), \code{"gsed1912"}, \code{"lf2206"}, \code{"sf2206"}, \code{"gcdg"},
and \code{"dutch"}. Since version 1.5.0, the \code{key = "gsed"}
selects the latest key starting with the string "gsed".
Use \code{key = ""} to use all item names,
which should only be done if there are no duplicate itemnames
in the itembank.}

\item{itembank}{A \code{data.frame} with at least two columns
named \code{item} and \code{tau}. By default, the function uses
\code{dscore::builtin_itembank}. If you specify your own \code{itembank},
then also specify the \code{transform} and \code{qd} arguments.}

\item{metric}{A string, either \code{"dscore"} (default) or
\code{"logit"}, signalling the metric in which ability is estimated.}

\item{prior_mean}{A string specifying where the mean of the
prior for the D-score calculation should come from. It could be
a column name in \code{data} (when you want your own prior for every row),
but normally this is one of the keywords \code{".dutch"}, \code{".gcdg"}
or \code{".phase1"}.
The default depends on the \code{key}. If \code{key == "dutch"} then
\code{prior_mean = ".dutch"}. The choice \code{prior_mean = ".dutch"}
calculates \code{prior_mean} from the Count model coded in
\code{dscore:::count_mu_dutch()}).
If \code{key} is #' \code{"gcdg"}, \code{"gsed1912"},
\code{"gsed2206"}, \code{"lf2206"} or \code{"sf2206"} then \code{prior_mean = ".gcdg"}.
This setting calculates an age-dependent prior mean internally according
to function \code{dscore:::count_mu_gcdg()}.
In other cases, \code{prior_mean = ".phase1"}
which uses the function \code{dscore:::count_mu_phase1()} or
\code{dscore:::count_mu_phase1_healthy()}.
Normally, you should not touch this parameter, but feel free to use
\code{prior_mean} to override the automatic choices.}

\item{prior_sd}{A string specifying a column name in \code{data}
with the standard deviation of the prior for the D-score calculation.
If not specified, the standard deviation is taken as 5 for every row.}

\item{transform}{Numeric vector, length 2, containing the intercept
and slope of the linear transform from the logit scale into the
the D-score scale. The default (\code{NULL}) searches \code{builtin_keys}
for intercept and slope values.}

\item{qp}{Numeric vector of equally spaced quadrature points.
This vector should span the range of all D-score or logit values.
The default (\code{NULL}) creates \code{seq(from, to, by)} searching the
arguments from \code{builtin_keys}.}

\item{population}{A string describing the population. Currently
supported are \code{"phase1"} (default), \code{"dutch"}, \code{"gcdg"}.}

\item{dec}{A vector of two integers specifying the number of
decimals for rounding the D-score and DAZ, respectively.
The default is \code{dec = c(2L, 3L)}.}

\item{relevance}{A numeric vector of length with the lower and
upper bounds of the relevance interval. The procedure calculates
a dynamic EAP for each item. If the difficulty level (tau) of the
next item is outside the relevance interval around EAP, the procedure
ignore the score on the item. The default is \code{c(-Inf, +Inf)} does not
ignore scores.}

\item{algorithm}{Computational method, for backward compatibility.
Either \code{"current"} (default) or \code{"1.8.7"} (deprecated).}
}
\value{
The \code{dscore()} function returns a \code{data.frame} with \code{nrow(data)} rows.
Optionally, the first block of columns can be specified by \code{prepend}
are copied from \code{data}. The second block consists of the
following columns:\tabular{ll}{
   Name \tab Label \cr
   \code{a} \tab Decimal age \cr
   \code{n} \tab Number of items with valid (0/1) data \cr
   \code{p} \tab Percentage of passed milestones \cr
   \code{d} \tab Ability estimate, mean of posterior \cr
   \code{sem} \tab Standard error of measurement, standard deviation of the posterior \cr
   \code{daz} \tab D-score corrected for age, calculated in Z-scale \cr
}


The \code{dscore_posterior()} function returns a data frame with
\code{nrow(data)} rows and \code{length(qp)} plus prepended columns with the
density at each quadrature point. A row vector representes the full
posterior ability distribution. If no valid responses are found,
\code{dscore_posterior()} returns the prior density. Versions prior to
1.8.5 returned a \code{matrix} (instead of a \code{data.frame}). Code that depends on
the result being a \code{matrix} may break and needs to be adapted.
}
\description{
The \code{dscore} package implements several tools needed to
calculate the D-score, a numerical score that measures
generic development in children.

The \code{dscore()} function estimates the D-score, a numeric score that
quantifies child development by one number.
}
\details{
The algorithm is based on the method by Bock and Mislevy (1982). The
method uses Bayes rule to update a prior ability into a posterior
ability.

The item names should correspond to the \code{"gsed"} lexicon.

A key is defined by the set of estimated item difficulties.\tabular{lrrrcl}{
   Key \tab Model \tab Quadrature \tab Instruments \tab Direct/Caregiver \tab Reference \cr
   \code{"dutch"} \tab \verb{75_0} \tab \code{-10:80} \tab 1 \tab direct \tab Van Buuren, 2014/2020 \cr
   \code{"gcdg"} \tab \verb{565_18} \tab \code{-10:100} \tab 14 \tab direct \tab Weber, 2019 \cr
   \code{"gsed1912"} \tab \verb{807_17} \tab \code{-10:100} \tab 20 \tab mixed \tab GSED Team, 2019 \cr
   \code{"gsed2206"} \tab \verb{818_17} \tab \code{-10:100} \tab 22 \tab mixed \tab GSED Team, 2022 \cr
   \code{"gsed2208"} \tab \verb{818_6} \tab \code{-10:100} \tab 22 \tab mixed \tab GSED Team, 2022 \cr
   \code{"gsed2212"} \tab \verb{818_6} \tab \code{-10:100} \tab 22 \tab mixed \tab GSED Team, 2022 \cr
   \code{"lf2206"} \tab \verb{155_0} \tab \code{-10:100} \tab 1 \tab direct \tab GSED Team, 2022 \cr
   \code{"sf2206"} \tab \verb{139_0} \tab \code{-10:100} \tab 1 \tab caregiver \tab GSED Team, 2022 \cr
}


As a general rule, one should only compare D-scores
that are calculated using the same key and the same
set of quadrature points. For calculating D-scores on new data,
the advice is to use the default, which currently links to
\code{"gsed2212"}.

The default starting prior is a mean calculated from a so-called
"Count model" that describes mean D-score as a function of age. The
Count models are stored as internal functions
\code{dscore:::count_mu_phase1()}, \code{dscore:::count_mu_gcdg()} and
\code{dscore:::count_mu_dutch()}. The spread of the starting prior
is 5 D-score points around this mean D-score, which corresponds to
approximately 1.5 to 2 times the normal spread of child of a given age. The
starting prior is thus somewhat informative for low numbers of
valid items, and uninformative for large number of items (say >10 items).
}
\note{
This study was supported by the Bill & Melinda Gates Foundation.
The contents are the sole responsibility of the authors and may not
necessarily represent the official views of the Bill & Melinda
Gates Foundation or other agencies that may have supported the
primary data studies used in the present study.

The authors wish to
recognize the principal investigators and their study team members
for their generous contribution of the data that made this tool
possible and the members of the Ki team who directly or indirectly
contributed to the study: Amina Abubakar, Claudia R. Lindgren
Alves, Orazio Attanasio, Maureen M. Black, Maria Caridad Araujo,
Susan M. Chang-Lopez, Gary L. Darmstadt, Bernice M. Doove, Wafaie
Fawzi, Lia C.H. Fernald, Günther Fink, Emanuela Galasso, Melissa
Gladstone, Sally M. Grantham-McGregor, Cristina Gutierrez de
Pineres, Pamela Jervis, Jena Derakhshani Hamadani, Charlotte
Hanlon, Simone M. Karam, Gillian Lancaster, Betzy Lozoff, Gareth
McCray, Jeffrey R Measelle, Girmay Medhin, Ana M. B. Menezes,
Lauren Pisani, Helen Pitchik, Muneera Rasheed, Lisy
Ratsifandrihamanana, Sarah Reynolds, Linda Richter, Marta
Rubio-Codina, Norbert Schady, Limbika Sengani, Chris Sudfeld,
Marcus Waldman, Susan P. Walker, Ann M. Weber and Aisha K.
Yousafzai.
}
\section{User functions}{

The available functions are:\tabular{ll}{
   Function \tab Description \cr
   \code{\link[=get_itemnames]{get_itemnames()}} \tab Extract item names from an itemtable \cr
   \code{\link[=order_itemnames]{order_itemnames()}} \tab Order item names \cr
   \code{\link[=sort_itemnames]{sort_itemnames()}} \tab Sort item names \cr
   \code{\link[=decompose_itemnames]{decompose_itemnames()}} \tab Get four components from itemname \cr
     \tab  \cr
   \code{\link[=get_itemtable]{get_itemtable()}} \tab Get a subset from the itemtable \cr
   \code{\link[=get_labels]{get_labels()}} \tab Get labels for items \cr
   \code{\link[=rename_gcdg_gsed]{rename_gcdg_gsed()}} \tab Rename gcdg into gsed lexicon \cr
     \tab  \cr
   \code{\link[=dscore]{dscore()}} \tab Estimate D-score and DAZ \cr
   \code{\link[=dscore_posterior]{dscore_posterior()}} \tab Calculate full posterior of D-score \cr
   \code{\link[=get_tau]{get_tau()}} \tab Get difficulty parameters from item bank \cr
     \tab  \cr
   \code{\link[=daz]{daz()}} \tab Transform to age-adjusted standardized D-score \cr
   \code{\link[=zad]{zad()}} \tab Inverse of \code{\link[=daz]{daz()}} \cr
   \code{\link[=get_reference]{get_reference()}} \tab Get D-score age-reference \cr
   \code{\link[=get_age_equivalent]{get_age_equivalent()}} \tab Translate difficulty to age \cr
}
}

\section{Built-in data}{

The package contains the following built-in data:\tabular{ll}{
   Data \tab Description \cr
   \code{\link[=builtin_itembank]{builtin_itembank()}} \tab A \code{data.frame} containing the difficulty estimates of items according to final Rasch models. \cr
   \code{\link[=builtin_itemtable]{builtin_itemtable()}} \tab A \code{data.frame} containing names and descriptions of items from 22 instruments. \cr
   \code{\link[=builtin_references]{builtin_references()}} \tab A \code{data.frame} with LMS reference values used to transform from D-score to DAZ, DAZ to D-score. \cr
   \code{\link[=milestones]{milestones()}} \tab A small demo dataset with PASS/FAIL responses from 27 preterms, measured at various ages between birth \cr
   and 2.5 years. \tab  \cr
}
}

\examples{
data <- data.frame(
  id = c(
    "Jane", "Martin", "ID-3", "No. 4", "Five", "6",
    NA_character_, as.character(8:10)
  ),
  age = rep(round(21 / 365.25, 4), 10),
  ddifmd001 = c(NA, NA, 0, 0, 0, 1, 0, 1, 1, 1),
  ddicmm029 = c(NA, NA, NA, 0, 1, 0, 1, 0, 1, 1),
  ddigmd053 = c(NA, 0, 0, 1, 0, 0, 1, 1, 0, 1)
)
items <- names(data)[3:5]

# third item is not part of default key
get_tau(items)

# calculate D-score
dscore(data)

# prepend id variable to output
dscore(data, prepend = "id")

# prepend all data
# dscore(data, prepend = colnames(data))

# calculate full posterior
p <- dscore_posterior(data)

# check that rows sum to 1
rowSums(p)

# plot posterior for row 7
barplot(as.matrix(p[7, 12:29]),
  names = 1:18,
  xlab = "D-score", ylab = "Density",
  main = "Full D-score posterior for measurement in row 7"
)
}
\references{
Jacobusse, G., S. van Buuren, and P.H. Verkerk. 2006. “An Interval Scale
for Development of Children Aged 0-2 Years.” \emph{Statistics in Medicine} 25 (13):
2272–83. \url{https://stefvanbuuren.name/publication/jacobusse-2006/}

Van Buuren S (2014). Growth charts of human development.
Stat Methods Med Res, 23(4), 346-368.
\url{https://stefvanbuuren.name/publication/van-buuren-2014-gc/}

Weber AM, Rubio-Codina M, Walker SP, van Buuren S, Eekhout I,
Grantham-McGregor S, Caridad Araujo M, Chang SM, Fernald LCH,
Hamadani JD, Hanlon A, Karam SM, Lozoff B, Ratsifandrihamanana L,
Richter L, Black MM (2019). The D-score: a metric for interpreting
the early development of infants and toddlers across global settings.
BMJ Global Health, BMJ Global Health 4: e001724.
\url{https://gh.bmj.com/content/bmjgh/4/6/e001724.full.pdf}.

GSED team (Maureen Black, Kieran Bromley, Vanessa Cavallera
(lead author), Jorge Cuartas, Tarun Dua (corresponding author),
Iris Eekhout, Gunther Fink, Melissa Gladstone, Katelyn Hepworth,
Magdalena Janus, Patricia Kariger, Gillian Lancaster, Dana McCoy,
Gareth McCray, Abbie Raikes, Marta Rubio-Codina, Stef van Buuren,
Marcus Waldman, Susan Walker and Ann Weber). 2019. “The Global Scale
for Early Development (GSED).” \emph{Early Childhood Matters}.
\url{https://earlychildhoodmatters.online/2019/the-global-scale-for-early-development-gsed/}

Bock DD, Mislevy RJ (1982).
Adaptive EAP Estimation of Ability in a Microcomputer Environment.
Applied Psychological Measurement, 6(4), 431-444.

Van Buuren S (2014). Growth charts of human development.
Stat Methods Med Res, 23(4), 346-368.
\url{https://stefvanbuuren.name/publication/van-buuren-2014-gc/}

Weber AM, Rubio-Codina M, Walker SP, van Buuren S, Eekhout I,
Grantham-McGregor S, Caridad Araujo M, Chang SM, Fernald LCH,
Hamadani JD, Hanlon A, Karam SM, Lozoff B, Ratsifandrihamanana L,
Richter L, Black MM (2019). The D-score: a metric for interpreting
the early development of infants and toddlers across global settings.
BMJ Global Health, BMJ Global Health 4: e001724.
\url{https://gh.bmj.com/content/bmjgh/4/6/e001724.full.pdf}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/d-score/dscore}
  \item \url{https://d-score.org/dscore/}
  \item \url{https://d-score.org/dbook1/}
  \item Report bugs at \url{https://github.com/d-score/dscore/issues}
}


\code{\link[=get_tau]{get_tau()}},
\code{\link[=builtin_itembank]{builtin_itembank()}}, \code{\link[=posterior]{posterior()}},
\code{\link[=builtin_references]{builtin_references()}}
}
\author{
\strong{Maintainer}: Stef van Buuren \email{stef.vanbuuren@tno.nl}

Authors:
\itemize{
  \item Iris Eekhout \email{iris.eekhout@tno.nl}
  \item Arjan Huizing \email{arjan.huizing@tno.nl}
}


Stef van Buuren, Iris Eekhout, Arjan Huizing (2022)
}
\keyword{package}
