% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_age_equivalent.R
\name{get_age_equivalent}
\alias{get_age_equivalent}
\title{Get age equivalents of items that have a difficulty estimate}
\usage{
get_age_equivalent(
  items,
  pct = c(10, 50, 90),
  key = NULL,
  transform = NULL,
  qp = NULL,
  itembank = dscore::builtin_itembank,
  population = NULL,
  xunit = c("decimal", "days", "months")
)
}
\arguments{
\item{items}{A character vector containing names of items to be
included into the D-score calculation. Milestone scores are coded
numerically as \code{1} (pass) and \code{0} (fail). By default,
D-score calculation is done on all items found in the data
that have a difficulty parameter under the specified \code{key}.}

\item{pct}{Numeric vector with requested percentiles (0-100). The
default is \code{pct = c(10, 50, 90)}.}

\item{key}{A string that selects a subset in the itembank that
makes up the key, the set of difficulty estimates from a fitted
Rasch model.
The built-in keys are:  \code{"gsed2212"} (default), \code{"gsed2208"} (deprecated),
\code{"gsed2206"} (deprecated), \code{"gsed1912"}, \code{"lf2206"}, \code{"sf2206"}, \code{"gcdg"},
and \code{"dutch"}. Since version 1.5.0, the \code{key = "gsed"}
selects the latest key starting with the string "gsed".
Use \code{key = ""} to use all item names,
which should only be done if there are no duplicate itemnames
in the itembank.}

\item{transform}{Numeric vector, length 2, containing the intercept
and slope of the linear transform from the logit scale into the
the D-score scale. The default (\code{NULL}) searches \code{builtin_keys}
for intercept and slope values.}

\item{qp}{Numeric vector of equally spaced quadrature points.
This vector should span the range of all D-score or logit values.
The default (\code{NULL}) creates \code{seq(from, to, by)} searching the
arguments from \code{builtin_keys}.}

\item{itembank}{A \code{data.frame} with at least two columns
named \code{item} and \code{tau}. By default, the function uses
\code{dscore::builtin_itembank}. If you specify your own \code{itembank},
then also specify the \code{transform} and \code{qd} arguments.}

\item{population}{A string describing the population. Currently
supported are \code{"phase1"} (default), \code{"dutch"}, \code{"gcdg"}.}

\item{xunit}{A string specifying the unit in which age is measured
(either \code{"decimal"}, \code{"days"} or \code{"months"}).
The default \code{"decimal"} corresponds to decimal age in years.}
}
\value{
Tibble with four columns: \code{item}, \code{d} (D-score),
\code{pct} (percentile), and \code{a} (age-equivalent, in \code{xunit} units).
}
\description{
This function calculates the ages at which a certain percent
in the reference population passes the items.
}
\details{
The function internally defines a scale factor given the key.
}
\examples{
get_age_equivalent(c("gpagmc018", "gtogmd026", "ddicmm050"))
}
