% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gettau.R
\name{gettau}
\alias{gettau}
\title{Obtain difficulty parameters from item bank}
\usage{
gettau(items, key = "gsed", itembank = dscore::builtin_itembank)
}
\arguments{
\item{items}{A character vector containing names of items to be 
included into the D-score calculation. Milestone scores are coded 
numerically as \code{1} (pass) and \code{0} (fail). By default, 
D-score calculation is done on all items found in the data 
that have a difficulty parameter under the specified \code{key}.}

\item{key}{A string that sets the key, the set of difficulty 
estimates from a fitted Rasch model.
The built-in keys are: \code{"gsed"} (default), \code{"gcdg"}, 
and \code{"dutch"}. See details.}

\item{itembank}{A \code{data.frame} with columns
\code{key}, \code{item}, \code{tau}, \code{instrument}, \code{domain}, 
\code{mode}, \code{number} and \code{label}.
The function uses \code{dscore::builtin_itembank} by 
default.}
}
\value{
A named vector with the difficulty estimate per item with
\code{length(items)} elements.
}
\description{
Searches the item bank for matching items, and returns the 
difficulty estimates. Matching is done by item name. Comparisons
are done in lower case.
}
\examples{
# difficulty levels in the GHAP lexicon
gettau(items = c("ddifmd001", "DDigmd052", "xyz"))

}
\seealso{
\code{\link{itembank}}, \code{\link{dscore}}
}
\author{
Stef van Buuren 2019
}
