% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gettau.R
\name{gettau}
\alias{gettau}
\title{Obtain difficulty parameters from item bank}
\usage{
gettau(items, lexicon = "gsed", itembank = dscore::itembank)
}
\arguments{
\item{items}{A character vector containing names of items to be 
included into the D-score calculation. Milestone scores are coded 
numerically as \code{1} (pass) and \code{0} (fail). By default, 
all items in the data are included that 1) appear in the specified 
lexicon, and 2) that have a difficulty parameter under the specified 
\code{model}.}

\item{lexicon}{Item naming scheme. The built-in itembank supports
lexicons \code{"gsed"} (default), \code{"gcdg"}, \code{"ghap"}, 
\code{"dutch1983"}, \code{"dutch1996"}, \code{"dutch2005"} 
and \code{"smocc"}.}

\item{itembank}{A \code{data.frame} that contains the item names 
(in various lexicons), the item label, and the item difficulty 
parameters \code{tau} (under various models). 
Lexicon column names start with \code{"lex_"}. 
The default uses the built-in \code{dscore::itembank} object.}
}
\value{
A named vector with the difficulty estimate per item with
\code{length(items)} elements.
}
\description{
Searches the item bank for matching items, and returns the 
difficulty estimates. Matching is done by item name. Comparisons
are done in lower case.
}
\examples{
# difficulty levels in the GHAP lexicon
gettau(items = c("ddifmd001", "DDigmd052", "xyz"))

# difficulty levels in the GHAP lexicon
gettau(items = c("GSFIXEYE", "gsmarm", "GSMARMR"), lexicon = "ghap")

# difficulty levels of same items in the SMOCC lexicon
gettau(items = c("v1430", "V1432", "v1432", "IdontExist"), lexicon = "smocc")

}
\seealso{
\code{\link{itembank}}, \code{\link{dscore}}
}
\author{
Stef van Buuren 2019
}
