% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dscore.R
\name{gettau}
\alias{gettau}
\title{Obtain difficulty parameters from item bank}
\usage{
gettau(items, itembank = dscore::itembank, lexicon = "lex.GHAP",
  check = TRUE, ...)
}
\arguments{
\item{items}{A character vector with item names in the 
chosen lexicon.}

\item{itembank}{A \code{data.frame} that contains the item names 
(in various lexicons), the item label, the item difficulty parameter
\code{tau}. Lexicon column names start with \code{"lex."}. 
The default \code{dscore::itembank} contains the 
difficulty levels as published in Van Buuren (2014).}

\item{lexicon}{A character string indicating the column in 
\code{itembank} used to match item names. It must be one of 
the lexicon columns, e.g., \code{lex.dutch1983}, 
\code{lex.dutch1996}, \code{lex.dutch2005}, \code{lex.smocc} 
or \code{lex.GHAP}. The default is \code{lex.GHAP}.}

\item{check}{Logical that indicates whether the lexicon name
should be checked. The default is \code{TRUE}.}

\item{\dots}{Additional arguments (ignored).}
}
\value{
A named vector with the difficulty estimate per item with
\code{length(items)} elements.
}
\description{
Searches the item bank for matching items, and returns the 
difficulty estimates. Matching is done by item name. Comparisons
are done in lower case.
}
\examples{
# difficulty levels in default GHAP lexicon
gettau(items = c("GSFIXEYE", "GSMARMR"))

# difficulty levels of same items in the SMOCC lexicon
gettau(items = c("v1430", "v1432"), lexicon = "lex.SMOCC")

# difficulty levels of same items in lex.dutch1996 lexicon
gettau(items = c("v1", "v51"), lexicon = "lex.dutch1996")
}
\author{
Stef van Buuren 2016
}
\seealso{
\code{\link{itembank}}, \code{\link{dscore}}
}

