% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{itembank}
\alias{itembank}
\title{Itembank containing difficulties per item}
\format{A \code{data.frame} with 84 rows and 11 variables:
\describe{
\item{ID.VWO1996}{Item name, VWO1996 version}
\item{ID.VWO2005}{Item name, VWO2005 version}
\item{ID.VWO1983}{Item name, VWO1983 version}
\item{ID.smock}{Item name, SMOCC data}
\item{occ}{Occasion number}
\item{labelNL}{Label (Dutch)}
\item{labelEN}{Label (English)}
\item{tau}{Difficulty level}
\item{m}{ }
\item{LR}{Left-right item scores}
\item{hot}{Item used to estimate D-score? (1=y, 0=no)}
}}
\usage{
itembank
}
\description{
A data frame with administrative information per item: 
item name (in various systems), item labels and difficulty.
}
\examples{
head(itembank)

}
\references{
Van Buuren S (2014). Growth charts of human development.
Stat Methods Med Res, 23(4), 346-368.
}
\seealso{
\code{\link{dscore}}, \code{\link{gettau}}
}
\keyword{datasets}

