% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dscore.R
\name{normalize}
\alias{normalize}
\title{Normalize distribution}
\usage{
normalize(d, qp)
}
\arguments{
\item{d}{A vector with \code{length(qp)} elements representing
the unscaled density at each quadrature point.}

\item{qp}{Vector of equally spaced quadrature points.}
}
\value{
A \code{vector} of \code{length(d)} elements with 
the prior density estimate at each quadature point.
}
\description{
Normalizes the distribution so that the total mass equals 1.
}
\examples{
# simple normalization examples
dscore:::normalize(c(5, 10, 5), qp = c(0, 1, 2))
dscore:::normalize(c(1, 5, 8, 5, 1), qp = 1:5)

# the sum is always equal to 1
sum(dscore:::normalize(rnorm(5), qp = 1:5))
}
