% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{posterior}
\alias{posterior}
\title{Calculate posterior for one item given score, difficulty and prior}
\usage{
posterior(score, delta, prior, qp)
}
\arguments{
\item{score}{A scalar value between 0 and m}

\item{delta}{A vector of length \code{m} with uncentralized difficulties from the Rasch model}

\item{prior}{vector of prior values on quadrature points \code{qp}}

\item{qp}{vector of equally spaced quadrature points}
}
\value{
A vector of length \code{length(prior)}
}
\description{
Calculate posterior for one item given score, difficulty and prior
}
\details{
This function assumes that the difficulties have been estimated by 
a binary Rasch model (e.g. by 
\code{sirt::rasch.pairwise.itemcluster()}) or - more generally - by 
a polytomous Rasch model (e.g. by \code{RUMM2030}), and 
transformed onto the 
appropriate scale. The response vector takes on values
\code{0:m}. The binary Rasch model assumes \code{m = 1}. The D-score
calculation assumes binary scores. 

Note: This function is intended to be used internally and 
does not check input arguments. Use \code{dscore()} to estimate
D-score from data.
}
\references{
Andrich D. Rasch Models for Measurement. Newbury Park: 
Sage Publications; 1988.

RUMM Laboratories. RUMM 2030. 
Rasch Unidimensional Measurement Models. Perth: 2015.
}
\seealso{
\code{\link{dscore}}, \code{\link{adp}}, 
\code{\link[sirt]{rasch.pairwise.itemcluster}}
}
\author{
Stef van Buuren 2016
}
