% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{posterior}
\alias{posterior}
\title{Calculate posterior for one item given score, difficulty and prior}
\usage{
posterior(score, tau, prior, qp)
}
\arguments{
\item{score}{Integer, either 0 (fail) and 1 (pass)}

\item{tau}{Numeric, difficulty parameter}

\item{prior}{Vector of prior values on quadrature points \code{qp}}

\item{qp}{vector of equally spaced quadrature points}
}
\value{
A vector of length \code{length(prior)}
}
\description{
This function assumes that the difficulties have been estimated by 
a binary Rasch model (e.g. by 
\code{sirt::rasch.pairwise.itemcluster()}).
}
\details{
Internal function. Not the be called directly.
}
\seealso{
\code{\link{dscore}}, 
\code{\link[sirt]{rasch.pairwise.itemcluster}}
}
\author{
Stef van Buuren 2018
}
